% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-methods.R
\name{as}
\alias{as}
\alias{coerce}
\alias{coerce,CellCycleMarkers,tbl_df-method}
\alias{coerce,CellTypeMarkers,tbl_df-method}
\alias{coerce,SingleCellExperiment,seurat-method}
\alias{coerce,seurat,SingleCellExperiment-method}
\alias{coerce,seurat,RangedSummarizedExperiment-method}
\alias{coerce,seurat,SummarizedExperiment-method}
\alias{coerce,SeuratMarkers,tbl_df-method}
\alias{coerce,SeuratMarkersPerCluster,tbl_df-method}
\title{Force an Object to Belong to a Class}
\description{
Coerce an object to a given class.
}
\section{\code{CellCycleMarkers} to \code{tbl_df}}{

S4 coercion support for creating a \code{tbl_df} from \code{CellCycleMarkers}.
}

\section{\code{CellTypeMarkers} to \code{tbl_df}}{

S4 coercion support for creating a \code{tbl_df} from \code{CellTypeMarkers}.
}

\section{\code{SingleCellExperiment} to \code{seurat}}{

Interally \code{Seurat::CreateSeuratObject} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \code{seurat} class object. Use
\code{convertGenesToSymbols} to convert gene IDs to names (symbols).
}

\section{\code{seurat} to \code{SingleCellExperiment}}{

S4 coercion support for creating a \code{SingleCellExperiment} from a \code{seurat}
class object. The \href{https://satijalab.org/seurat/faq}{Seurat FAQ page}
explains the \code{seurat} S4 class structure in detail. Internally, this method
improves the basic \code{Seurat::as.SingleCellExperiment} S3 coercion method,
including the \code{object@scale.data} matrix, and will keep track of stashed
\code{rowRanges} and \code{metadata} if the \code{seurat} object was originally created
from a \code{SingleCellExperiment} (i.e. from the bcbioSingleCell package).
}

\section{\code{seurat} to \code{RangedSummarizedExperiment}}{

S4 coercion support for creating a \code{RangedSummarizedExperiment} from a
\code{seurat} class object.
}

\section{\code{seurat} to \code{SummarizedExperiment}}{

S4 coercion support for creating a \code{SummarizedExperiment} from a \code{seurat}
class object.
}

\section{\code{SeuratMarkers} to \code{tbl_df}}{

S4 coercion support for creating a \code{tbl_df} from a \code{Markers} object.
}

\section{\code{SeuratMarkersPerCluster} to \code{tbl_df}}{

S4 coercion support for creating a \code{tbl_df} from \code{SeuratMarkersPerCluster}.
}

\section{Description}{

    \code{as(object)}
      returns the version of this object coerced to be the given
      \code{Class}.  When used in the replacement form on the left of
      an assignment, the portion of the object corresponding to
      \code{Class} is replaced by \code{value}.

      The operation of \code{as()} in either form depends on the
      definition of coerce methods.  Methods are defined automatically
      when the two classes are related by inheritance; that is, when
      one of the classes is a subclass of the other.

      Coerce methods are also predefined for basic classes (including all
      the types of vectors, functions and a few others).

      Beyond these two sources of methods, further methods are defined
      by calls to the \code{\link{setAs}} function.  See that
      documentation also for details of how coerce methods work. Use
      \code{showMethods(coerce)} for a list of all currently defined methods, as in the
      example below.

 
  
}

\section{Basic Coercion Methods}{

  Methods are pre-defined for coercing any object to one of the basic
  datatypes.  For example, \code{as(x, "numeric")} uses the existing
  \code{as.numeric} function.  These and all other existing methods
  can be listed as shown in the example.

}

\examples{
data(sce, package = "basejump")
data(seurat_small)

## SingleCellExperiment to seurat ====
object <- sce
print(object)
x <- as(object, "seurat")
class(x)
print(x)

## seurat to SingleCellExperiment ====
x <- as(seurat_small, "SingleCellExperiment")
print(x)
}
\references{
Chambers, John M. (2016)
 \emph{Extending R},
  Chapman & Hall.
(Chapters 9 and 10.)
}
\seealso{
\itemize{
\item \code{Seurat::CreateSeuratObject}.
\item \code{Seurat::Convert}.
}
}
