% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/as.Seurat-methods.R
\name{as.Seurat}
\alias{as.Seurat}
\alias{as.Seurat,SingleCellExperiment-method}
\title{Coerce object to Seurat}
\usage{
as.Seurat(x, ...)

\S4method{as.Seurat}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Seurat}.
}
\description{
Coerce object to Seurat
}
\details{
Note that \code{Seurat::as.Seurat()} method requires \code{logcounts} to be defined
in \code{assays()}, so we're using \code{CreateSeuratObject()} here instead.
}
\note{
Updated 2023-10-04.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment to Seurat ====
from <- SingleCellExperiment_Seurat
to <- as.Seurat(from)
class(to)
print(to)
}
\seealso{
\itemize{
\item \code{Seurat::CreateSeuratObject()}.
\item \code{Seurat::as.Seurat()}.
}
}
