% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-cell_data_set-methods.R
\name{base-cell_data_set}
\alias{base-cell_data_set}
\alias{Gene2Symbol,cell_data_set-method}
\alias{logcounts,cell_data_set-method}
\alias{metrics,cell_data_set-method}
\alias{normcounts,cell_data_set-method}
\alias{sizeFactors,cell_data_set-method}
\alias{sizeFactors<-,cell_data_set,ANY-method}
\title{Extend base S4 methods for \code{cell_data_set} class}
\usage{
\S4method{Gene2Symbol}{cell_data_set}(object, ...)

\S4method{logcounts}{cell_data_set}(object, verbose = FALSE)

\S4method{metrics}{cell_data_set}(object, return = c("tbl_df", "DataFrame"))

\S4method{normcounts}{cell_data_set}(object)

\S4method{sizeFactors}{cell_data_set}(object)

\S4method{sizeFactors}{cell_data_set,ANY}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{value}{Value to assign.}
}
\value{
Varies, depending on the generic.
}
\description{
Extend base S4 methods for \code{cell_data_set} class
}
\note{
Updated 2020-01-30.
}
\keyword{internal}
