% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_cycle_markers}
\alias{cell_cycle_markers}
\alias{cellCycleMarkers}
\title{Cell-cycle markers}
\format{An object of class \code{list} of length 3.}
\usage{
cell_cycle_markers
}
\description{
Cell-cycle markers
}
\examples{
data(cell_cycle_markers)
cell_cycle_markers
}
\keyword{datasets}
