% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_cycle_markers_list}
\alias{cell_cycle_markers_list}
\title{Cell-cycle markers list}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(cell_cycle_markers_list)
}
\value{
\code{list}.
}
\description{
Cell-cycle markers list
}
\note{
Updated 2020-01-30.
}
\examples{
data(cell_cycle_markers_list)
names(cell_cycle_markers_list)
}
\keyword{datasets}
