% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_data_set}
\alias{cell_data_set}
\title{Cell data set}
\format{An object of class \code{cell_data_set} with 200 rows and 200 columns.}
\usage{
data(cell_data_set)
}
\value{
\code{cell_data_set}.
}
\description{
Example monocle3 \code{cell_data_set} object.
}
\note{
Updated 2020-01-30.
}
\examples{
if (requireNamespace("monocle3", quietly = TRUE)) {
    data(cell_data_set)
    class(cell_data_set)
    cell_data_set
}
}
\keyword{datasets}
