% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_markers}
\alias{cell_type_markers}
\alias{cellTypeMarkers}
\title{Cell-type markers}
\format{An object of class \code{list} of length 3.}
\usage{
cell_type_markers
}
\description{
Cell-type markers
}
\examples{
data(cell_type_markers)
cell_type_markers
}
\keyword{datasets}
