% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{\code{CellCycleMarkers} Class}
\value{
\code{CellCycleMarkers}.
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=attributes]{attributes()}}.
}
\section{Slots}{

\describe{
\item{\code{organism}}{\code{string}. Full Latin organism name.}

\item{\code{ensemblRelease}}{\code{scalar integer}. Ensembl release version.}

\item{\code{version}}{\code{package_version}.}

\item{\code{date}}{\code{Date}. Date the object was saved.}
}}

\seealso{
Other S4 Classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
