% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellCycleMarkersList}
\alias{cellCycleMarkersList}
\title{Cell-cycle markers list}
\format{An object of class \code{list} of length 3.}
\usage{
cellCycleMarkersList
}
\description{
Cell-cycle markers list
}
\examples{
data(cellCycleMarkersList)
names(cellCycleMarkersList)
}
\keyword{datasets}
