% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMarkers-methods.R
\name{CellMarkers}
\alias{CellMarkers}
\alias{CellCycleMarkers}
\alias{CellTypeMarkers}
\title{Cell markers}
\usage{
CellCycleMarkers(object, gene2symbol)

CellTypeMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}
}
\description{
Cell markers
}
\section{Functions}{
\itemize{
\item \code{CellCycleMarkers}: Cell-cycle markers.

\item \code{CellTypeMarkers}: Cell-type markers.
}}

\note{
Updated 2019-07-31.
}
