% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellTypeMarkers-class}
\alias{CellTypeMarkers-class}
\title{\code{CellTypeMarkers} Class}
\value{
\code{CellTypeMarkers}.
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=metadata]{metadata()}}.
}
\seealso{
Other S4 Classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
