% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{CellTypeMarkers}
\alias{CellTypeMarkers}
\title{Cell-Type Markers}
\usage{
CellTypeMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings.}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=metadata]{metadata()}}.
}
\seealso{
\code{\link[=CellTypeMarkers]{CellTypeMarkers()}}.

Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{KnownSeuratMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
