% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellTypeMarkersList}
\alias{cellTypeMarkersList}
\title{Cell-type markers list}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(cellTypeMarkersList)
}
\value{
\code{list}.
}
\description{
Cell-type markers list
}
\note{
Updated 2021-03-03.
}
\examples{
data(cellTypeMarkersList)
names(cellTypeMarkersList)
}
\keyword{datasets}
