% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{coerce,Seurat,SingleCellExperiment-method}
\alias{coerce,Seurat,RangedSummarizedExperiment-method}
\alias{coerce,Seurat,SummarizedExperiment-method}
\alias{coerce,SingleCellExperiment,Seurat-method}
\title{Force an object to belong to a class}
\description{
Force an object to belong to a class
}
\note{
Updated 2023-08-16.
}
\examples{
data(Seurat, SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment to Seurat ====
from <- SingleCellExperiment_Seurat
to <- as(from, "Seurat")
class(to)
print(to)

## Seurat to SingleCellExperiment ====
from <- Seurat
to <- as(from, "SingleCellExperiment")
class(to)
print(to)
}
