% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-counts-methods.R, R/base-cpm-methods.R,
%   R/base-logcounts-methods.R, R/base-normcounts-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,Seurat-method}
\alias{cpm,Seurat-method}
\alias{logcounts}
\alias{logcounts,Seurat-method}
\alias{logcounts,cell_data_set-method}
\alias{normcounts}
\alias{normcounts,Seurat-method}
\alias{normcounts,cell_data_set-method}
\title{Named assay fields}
\usage{
counts(object, ...)

\S4method{counts}{Seurat}(object, assay = NULL)

cpm(object, ...)

\S4method{cpm}{Seurat}(object, assay = NULL)

logcounts(object, ...)

\S4method{logcounts}{Seurat}(object, assay = NULL)

\S4method{logcounts}{cell_data_set}(object)

normcounts(object, ...)

\S4method{normcounts}{Seurat}(object, assay = NULL)

\S4method{normcounts}{cell_data_set}(object)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{assay}{Name of assay to pull data from}

\item{...}{Additional arguments.}
}
\value{
Each method returns a matrix from the correspondingly named field in the \code{assays} slot.
}
\description{
Convenience methods to get or set named assay fields.
}
\details{
These are wrapper methods for getting or setting \code{assay(object, i=X, ...)} where \code{X} is the name of the method.
For example, \code{counts} will get or set \code{X="counts"}.
This provide some convenience for users as well as encouraging standardization of naming across packages.

Our suggested interpretation of the fields are as follows:
\describe{
\item{\code{counts}:}{Raw count data, e.g., number of reads or transcripts.}
\item{\code{normcounts}:}{Normalized values on the same scale as the original counts.
For example, counts divided by cell-specific size factors that are centred at unity.}
\item{\code{logcounts}:}{Log-transformed counts or count-like values.
In most cases, this will be defined as log-transformed \code{normcounts}, e.g., using log base 2 and a pseudo-count of 1.}
\item{\code{cpm}:}{Counts-per-million.
This is the read count for each gene in each cell, divided by the library size of each cell in millions.}
\item{\code{tpm}:}{Transcripts-per-million.
This is the number of transcripts for each gene in each cell, divided by the total number of transcripts in that cell (in millions).}
\item{\code{weights}:}{A matrix of weights, e.g., observational weights to be used in differential expression analysis.}
}
}
\examples{
example(SingleCellExperiment, echo=FALSE) # Using the class example
counts(sce) <- matrix(rnorm(nrow(sce)*ncol(sce)), ncol=ncol(sce))
dim(counts(sce))

# One possible way of computing normalized "counts"
sf <- 2^rnorm(ncol(sce))
sf <- sf/mean(sf)
normcounts(sce) <- t(t(counts(sce))/sf)
dim(normcounts(sce))

# One possible way of computing log-counts
logcounts(sce) <- log2(normcounts(sce)+1)
dim(normcounts(sce))
}
\seealso{
\itemize{
\item \code{Seurat::GetAssayData()}.
\item \code{Seurat::NormalizeData()}.
\item \code{Seurat::ScaleData()}.
}

\itemize{
\item \code{SingleCellExperiment::cpm()}.
\item \code{edgeR::cpm()}.
\item \code{scater::calculateCPM()}.
}

\itemize{
\item \code{scater::normalizeCounts()}.
}
}
