% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{knownMarkers}
\alias{knownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{readCellTypeMarkers}
\alias{sanitizeMarkers}
\title{Defunct functions}
\usage{
knownMarkers(...)

knownMarkersDetected(...)

plotKnownMarkersDetected(...)

readCellTypeMarkers(...)

sanitizeMarkers(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\keyword{internal}
