% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSizeFactors-methods.R
\docType{methods}
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\alias{estimateSizeFactors,DelayedArray-method}
\alias{estimateSizeFactors,SingleCellExperiment-method}
\alias{estimateSizeFactors,cell_data_set-method}
\title{Estimate size factors}
\usage{
estimateSizeFactors(object, ...)

\S4method{estimateSizeFactors}{DelayedArray}(object,
  type = c("mean-ratio", "geometric-mean-ratio",
  "log-geometric-mean-ratio", "deseq2-median-ratio"))

\S4method{estimateSizeFactors}{SingleCellExperiment}(object,
  colname = "sizeFactor")

\S4method{estimateSizeFactors}{cell_data_set}(object,
  colname = "Size_Factor")
}
\arguments{
\item{type}{\code{character(1)}.
Method for estimation:\preformatted{libSize <- colSums(counts(object))
}
\itemize{
\item \code{mean-ratio}: From scater.\preformatted{libSize / mean(libSize)
}
\item \code{geometric-mean-ratio}: From monocle3.\preformatted{libSize / geometricMean(libSize)
}
\item \code{mean-geometric-mean-log-total}: From monocle3.\preformatted{log(libSize) / geometricMean(log(libSize))
}
\item \code{median-ratio}: From DESeq2.
The size factor is a median ratio of the sample over a "pseudosample":
for each feature (i.e. gene), the geometric mean of all samples.
}}

\item{colname}{\code{character(1)}.
Column name where to define size factor values in
\code{\link[SummarizedExperiment:colData]{colData()}}.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{Matrix}: Named \code{numeric}, containing size factor values.
\item \code{SingleCellExperiment}: Modified object, containing size factor values
as \code{sizeFactor} column in \code{\link[SummarizedExperiment:colData]{colData()}}.
\item \code{cell_data_set}: Modified object, containing size factor values
as \code{Size_Factor} column in \code{\link[SummarizedExperiment:colData]{colData()}}.
}
}
\description{
Estimate size factors
}
\note{
Updated 2019-08-04.
}
\examples{
data(
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
sizeFactors(object)
object <- estimateSizeFactors(object)
sizeFactors(object)

## cell_data_set ====
object <- cell_data_set
object <- estimateSizeFactors(cell_data_set)
sizeFactors(object)
}
\seealso{
\itemize{
\item \code{scater::librarySizeFactors()}.
\item \code{monocle3::estimate_size_factors()}.
\item \code{monocle3:::estimate_sf_sparse()}.
\item \code{DESeq2::estimateSizeFactors()}.
\item \code{DESeq2::estimateSizeFactorsForMatrix().}
}
}
