% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fetchData-methods.R
\docType{methods}
\name{fetchGeneData}
\alias{fetchGeneData}
\alias{fetchReducedDimData}
\alias{fetchReducedDimExpressionData}
\alias{fetchData}
\alias{fetchGeneData,SingleCellExperiment-method}
\alias{fetchReducedDimData,SingleCellExperiment-method}
\alias{fetchReducedDimExpressionData,SingleCellExperiment-method}
\title{Fetch Data Functions}
\usage{
fetchGeneData(object, ...)

fetchReducedDimData(object, ...)

fetchReducedDimExpressionData(object, ...)

\S4method{fetchGeneData}{SingleCellExperiment}(object, genes,
  gene2symbol = FALSE)

\S4method{fetchReducedDimData}{SingleCellExperiment}(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups)

\S4method{fetchReducedDimExpressionData}{SingleCellExperiment}(object,
  genes, reducedDim)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{gene2symbol}{\code{data.frame}. Gene-to-symbol mappings. Columns must
contain \code{geneID} and \code{geneName}.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{dimsUse}{\code{integer}. Vector of length 2 that denotes the columns from
the reduced dimension matrix to use for \code{centerX} and \code{centerY} column
calculations. Defaults the first and second dimensions.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\itemize{
\item \code{fetchGeneData()}: \code{matrix}.
\item \code{fetchTSNEExpressionData()}: \code{data.frame} containing t-SNE coordinates,
sample metadata, and aggregate marker expression values (\code{mean}, \code{median},
and \code{sum}).
\item Other functions: \code{data.frame}, containing metrics.
}
}
\description{
Fetch Data Functions
}
\note{
Some of these functions require "\code{ident}" to be defined in \code{\link[=colData]{colData()}}.
For t-SNE, UMAP, and PCA, the reduced dimensions must be defined in the
\code{\link[=reducedDims]{reducedDims()}} slot.
}
\examples{
object <- sce_small
genes <- head(rownames(object))
glimpse(genes)

# Genes
x <- fetchGeneData(object, genes = genes)
glimpse(x)

# Reduced dim coordinates
x <- fetchReducedDimExpressionData(
    object = object,
    genes = genes,
    reducedDim = "TSNE"
)
glimpse(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Data Functions}
