% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findAllMarkers-methods.R
\docType{methods}
\name{findAllMarkers}
\alias{findAllMarkers}
\title{Find All Cluster-Specific Markers}
\usage{
findAllMarkers(object, ...)

\S4method{findAllMarkers}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list}
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find All Cluster-Specific Markers
}
\examples{
object <- sce_small

# edgeR
x <- findAllMarkers(object, caller = "edgeR")

# DESeq2
x <- findAllMarkers(object, caller = "DESeq2")
}
\seealso{
Other Differential Expression Functions: \code{\link{diffExp}},
  \code{\link{runZinbwave}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
