% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findMarkers-methods.R
\docType{methods}
\name{findMarkers}
\alias{findMarkers}
\alias{findMarkers,SingleCellExperiment-method}
\title{Find Cluster-Specific Marker Genes}
\usage{
findMarkers(object, ...)

\S4method{findMarkers}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find Cluster-Specific Marker Genes
}
\examples{
x <- suppressMessages(
    findMarkers(sce_small, caller = "edgeR")
)
class(x)
lapply(x, class)
}
\seealso{
Other Differential Expression Functions: \code{\link{diffExp}},
  \code{\link{runZinbwave}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
