% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene2Symbol-methods.R
\docType{methods}
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,Seurat-method}
\alias{Gene2Symbol,cell_data_set-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{Seurat}(object, ...)

\S4method{Gene2Symbol}{cell_data_set}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
Contains a \code{DataFrame} with \code{geneID} and \code{geneName} columns.
}
\description{
Gene-to-symbol mappings
}
\note{
Updated 2019-08-02.
}
\section{Genome metadata}{



We recommend slotting \code{organism}, \code{genomeBuild}, and \code{ensemblRelease} into
\code{\link[S4Vectors:metadata]{metadata()}}.

}

\examples{
data(
    Seurat,
    cell_data_set,
    package = "acidtest"
)

## Seurat ====
object <- Seurat
x <- Gene2Symbol(object)
head(x)
table(x)

## cell_data_set ====
object <- cell_data_set
x <- Gene2Symbol(object)
head(x)
table(x)
}
\seealso{
\code{\link[=makeGene2Symbol]{makeGene2Symbol()}}.
}
