% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbwave.R
\name{hasZinbwave}
\alias{hasZinbwave}
\title{Does Object Contain ZINB-WaVE Weights}
\usage{
hasZinbwave(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{logical}.
}
\description{
Does Object Contain ZINB-WaVE Weights
}
\examples{
hasZinbwave(sce_small)
}
\seealso{
Other Zero Count Inflation Functions: \code{\link{runZinbwave}}
}
\author{
Michael Steinbaugh
}
\concept{Zero Count Inflation Functions}
