% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ident-methods.R
\docType{methods}
\name{ident}
\alias{ident}
\alias{ident,SingleCellExperiment-method}
\title{Cluster Identity}
\usage{
ident(object, ...)

\S4method{ident}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster Identity
}
\examples{
ident(sce_small)
}
\author{
Michael Steinbaugh
}
\concept{Marker Functions}
