% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMarkers.R
\name{importMarkers}
\alias{importMarkers}
\alias{importCellCycleMarkersFromGoogle}
\alias{importCellTypeMarkersFromGoogle}
\title{Import Markers from Google}
\usage{
importCellCycleMarkersFromGoogle(gs, ws, gene2symbol)

importCellTypeMarkersFromGoogle(gs, ws, gene2symbol)
}
\arguments{
\item{gs}{\code{character(1)}.
Google Sheets \code{sheet_key} identifier, which can be located with the
\code{\link[googlesheets:gs_ls]{googlesheets::gs_ls()}} function.}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings.}
}
\value{
\code{CellCycleMarkers} or \code{CellTypeMarkers}.
}
\description{
Currently designed for internal use by the pointillism package.
}
\section{Functions}{
\itemize{
\item \code{importCellCycleMarkersFromGoogle}: Import Cell-Cycle Markers from Google

\item \code{importCellTypeMarkersFromGoogle}: Cell-type markers.
}}

\note{
Google Sheets requires OAuth authentication.
}
\examples{
## Must be interactive.
if (isTRUE(interactive())) {
    googlesheets::gs_ls()

    ## Gene-to-symbol mappings.
    gene2symbol <- makeGene2SymbolFromEnsembl("Homo sapiens")

    ## Cell-cycle markers.
    x <- importCellCycleMarkersFromGoogle(
        gs = "1qA5ktYeimNGpZF1UPSQZATbpzEqgyxN6daoMOjv6YYw",
        ws = "Homo_sapiens",
        gene2symbol = gene2symbol
    )

    ## Cell-type markers.
    x <- importCellTypeMarkersFromGoogle(
        gs = "1vGNU2CCxpaoTCLvzOxK1hf5gjULrf2-CpgCp9bOfGJ0",
        ws = "Homo_sapiens",
        gene2symbol = gene2symbol
    )
}
}
\seealso{
\itemize{
\item \code{googlesheets::gs_key()}.
\item \code{googlesheets::gs_read()}.
}
}
