% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_markers_detected_small}
\alias{known_markers_detected_small}
\title{Known Markers Detected Example}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 19 columns.}
\usage{
known_markers_detected_small
}
\description{
Known Markers Detected Example
}
\examples{
glimpse(known_markers_detected_small)
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{known_markers_small}},
  \code{\link{sce_small}}, \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
