% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_markers_small}
\alias{known_markers_small}
\title{Known Markers Example}
\format{An object of class \code{KnownSeuratMarkers} with 0 rows and 9 columns.}
\usage{
known_markers_small
}
\description{
Known Markers Example
}
\examples{
data(known_markers_small)
known_markers_small
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
