% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KnownMarkers-class}
\alias{KnownMarkers-class}
\title{Known markers}
\value{
\code{KnownMarkers}.
}
\description{
Class containing known markers detected.
}
\details{
Results are grouped by \code{cellType} column and arranged by adjusted \emph{P} value
(\code{padj}).
}
\note{
Updated 2019-07-31.
}
