% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/knownMarkersDetected-methods.R
\name{knownMarkersDetected}
\alias{knownMarkersDetected}
\title{Known Markers Detected}
\usage{
knownMarkersDetected(all, known, ...)
}
\arguments{
\item{all}{\code{SeuratMarkers}.}

\item{known}{\code{CellTypeMarkers}. Grouped by \code{cellType} column. Known markers
\code{data.frame} imported by \code{\link[=readCellTypeMarkers]{readCellTypeMarkers()}} or pulled from internal
\link{cell_cycle_markers} data.}

\item{...}{Additional arguments.}

\item{alpha}{\code{scalar numeric}. Alpha cutoff (adjusted P value; false
discovery rate).}

\item{filterPromiscuous}{\code{boolean}. Remove genes with poor specificity, that
are present in as many clusters as defined by \code{promiscuousCutoff}.}

\item{promiscuousCutoff}{\code{scalar integer}. Minimum number of clusters
required to consider a gene marker promiscuous.}
}
\value{
\code{DataFrame}.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
x <- knownMarkersDetected(
    all = all_markers_small,
    known = known_markers_small
)
head(x)
}
\seealso{
Other Marker Functions: \code{\link{CellTypeMarkers}},
  \code{\link{SeuratMarkers}}, \code{\link{clusterID}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Marker Functions}
