% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KnownSeuratMarkers-class}
\alias{KnownSeuratMarkers-class}
\title{\code{KnownSeuratMarkers} Class}
\description{
Class containing essential elements for analysis of known, cell-type specific
marker genes.
}
\seealso{
\code{\link[=knownMarkers]{knownMarkers()}}.

Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 classes}
