% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-logcounts-methods.R
\name{logcounts}
\alias{logcounts}
\title{Log normalized counts}
\usage{
logcounts(object, ...)
}
\value{
\code{sparseMatrix}.
}
\description{
Log normalized counts
}
\note{
Updated 2019-08-05.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ===
object <- SingleCellExperiment
object <- normalize(object)
logcounts <- logcounts(object)
class(logcounts)

## Seurat ====
object <- Seurat
logcounts <- logcounts(object)
class(logcounts)

## cell_data_set ====
object <- cell_data_set
object <- normalize(object)
logcounts <- logcounts(object)
class(logcounts)
}
\seealso{
\itemize{
\item \code{\link[=normalize]{normalize()}}.
\item \code{scater::normalizeSCE()}.
\item \code{Seurat::NormalizeData()}.
\item \code{monocle3::prepare_cds()}.
}
}
