% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MarkersPerCluster-class}
\alias{MarkersPerCluster-class}
\title{Single-Cell Markers per Cluster}
\value{
Results are split per \code{cluster} and arranged by adjusted \emph{P} value
(\code{padj}).
}
\description{
Class containing essential elements for marker gene analysis.
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{Markers-class}}
}
\concept{S4 classes}
