% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-methods.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics,cell_data_set-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{cell_data_set}(object, return = c("tibble",
  "DataFrame"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"tibble"}: \code{grouped_df}.
Grouped by \code{sampleID} column.
\item \code{"DataFrame"}: \code{DataFrame}.
Row names are identical to the column names of the object, like
\code{\link[SummarizedExperiment:colData]{colData()}}.
}
}
\description{
Quality control metrics
}
\details{
\code{\link[=metrics]{metrics()}} takes data stored in \code{\link[SummarizedExperiment:colData]{colData()}}
and consistently returns a tibble grouped by sample by default (\code{sampleID}).
It always returns \code{sampleName} and \code{interestingGroups} columns, even when
these columns are not defined in \code{colData}. This is designed to integrate
with plotting functions that use ggplot2 internally.
}
\note{
Updated 2019-08-02.
}
\section{Methods (by class)}{

\itemize{
\item \code{SummarizedExperiment}: Metrics are sample level. \code{sampleID} column corresponds
to \code{colnames}.

\item \code{SingleCellExperiment}: Metrics are cell level. \code{cellID} column corresponds to
\code{colnames}. Tibble is returned grouped by sample (\code{sampleID} column).
}
}

\examples{
data(
    cell_data_set,
    package = "acidtest"
)

## cell_data_set ====
object <- cell_data_set
x <- metrics(object)
print(x)

x <- metricsPerSample(object, fun = "mean")
print(x)
}
