% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-normcounts-methods.R
\name{normcounts}
\alias{normcounts}
\title{Size factor adjusted normalized counts}
\usage{
normcounts(object, ...)
}
\value{
\code{sparseMatrix}.
}
\description{
Size factor adjusted normalized counts
}
\note{
Updated 2019-08-05.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ===
object <- SingleCellExperiment
object <- normalize(object)
normcounts <- normcounts(object)
class(normcounts)

## Seurat ====
object <- Seurat
normcounts <- normcounts(object)
class(normcounts)

## cell_data_set ====
object <- cell_data_set
object <- normalize(object)
normcounts <- normcounts(object)
class(normcounts)
}
\seealso{
\itemize{
\item \code{\link[=normalize]{normalize()}}.
\item \code{scater::normalizeSCE()}.
\item \code{scater::normalizeCounts()}.
\item \code{Seurat::NormalizeData()}.
\item \code{monocle3::prepare_cds()}.
\item \code{monocle3::normalized_counts()}.
}
}
