% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters}
\arguments{
\item{alpha}{\code{numeric(1)}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}

\item{reduction}{\code{character(1)}.
Name of dimension reduction matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes UMAP, TSNE, PCA, typically, for example.}
}
\value{
No value.
}
\description{
Parameters
}
\keyword{internal}
