% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotCellTypesPerCluster-methods.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,SingleCellExperiment-method}
\alias{plotCellTypesPerCluster,seurat-method}
\title{Plot Cell Types per Cluster}
\usage{
plotCellTypesPerCluster(object, ...)

\S4method{plotCellTypesPerCluster}{SingleCellExperiment}(object, markers,
  min = 1L, max = Inf, reducedDim = 1L, expression = c("mean",
  "median", "sum"), headerLevel = 2L)

\S4method{plotCellTypesPerCluster}{seurat}(object, markers, min = 1L,
  max = Inf, reducedDim = 1L, expression = c("mean", "median",
  "sum"), headerLevel = 2L)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}

\item{markers}{\code{KnownSeuratMarkers}.}

\item{min}{\code{scalar numeric}. Recommended minimum value cutoff.}

\item{max}{\code{scalar numeric}. Recommended maximum value cutoff.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{\link[=match.arg]{match.arg()}} and
defaults to the first argument in the \code{character} vector.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{list}.
}
\description{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
data(seurat_small, known_markers_small)
plotCellTypesPerCluster(
    object = seurat_small,
    markers = known_markers_small
)
}
\seealso{
Other Plot Functions: \code{\link{plotDot}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkers}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
