% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R, R/plotDots-methods.R,
%   R/plotViolin-methods.R
\name{plotCounts}
\alias{plotCounts}
\alias{plotDots}
\alias{plotViolin}
\alias{plotCounts,SingleCellExperiment-method}
\alias{plotCounts,Seurat-method}
\alias{plotDots,SingleCellExperiment-method}
\alias{plotDots,Seurat-method}
\alias{plotViolin,SingleCellExperiment-method}
\alias{plotViolin,Seurat-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{SingleCellExperiment}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  geom = c("violin", "dot"),
  perSample = TRUE,
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotCounts}{Seurat}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  geom = c("violin", "dot"),
  perSample = TRUE,
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)

plotDots(object, ...)

\S4method{plotDots}{SingleCellExperiment}(
  object,
  genes,
  perSample = TRUE,
  colMin = -2.5,
  colMax = 2.5,
  dotMin = 0L,
  dotScale = 6L,
  color = getOption("acid.continuous.color", default =
    ggplot2::scale_color_gradient2(low = "orange", mid = "gray75", high = "purple",
    midpoint = 0L)),
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotDots}{Seurat}(
  object,
  genes,
  perSample = TRUE,
  colMin = -2.5,
  colMax = 2.5,
  dotMin = 0L,
  dotScale = 6L,
  color = getOption("acid.continuous.color", default =
    ggplot2::scale_color_gradient2(low = "orange", mid = "gray75", high = "purple",
    midpoint = 0L)),
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)

plotViolin(object, ...)

\S4method{plotViolin}{SingleCellExperiment}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption("acid.discrete.color", default =
    acidplots::scale_color_synesthesia_d()),
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)

\S4method{plotViolin}{Seurat}(
  object,
  genes,
  assay = c("logcounts", "normcounts"),
  perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption("acid.discrete.color", default =
    acidplots::scale_color_synesthesia_d()),
  legend = getOption("acid.legend", default = TRUE),
  title = NULL
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} to pick the type.
Currently supports \code{"dot"} and \code{"violin"}.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{title}{\code{character(1)}.
Title.}

\item{colMin}{\code{numeric(1)}.
Minimum scaled average expression threshold. Everything smaller will be
set to this.}

\item{colMax}{\code{numeric(1)}.
Maximum scaled average expression threshold. Everything larger will be set
to this.}

\item{dotMin}{\code{numeric(1)}.
The fraction of cells at which to draw the smallest dot. All cell groups
with less than this expressing the given gene will have no dot drawn.}

\item{dotScale}{\code{numeric(1)}.
Scale the size of the points, similar to \code{cex}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize genes on a dot or violin plot.
}
\note{
Dot geom currently only supports logcounts.

Updated 2020-02-21.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat

## Plotting with either gene IDs or gene names (symbols) works.
genes <- head(rownames(object), n = 4L)
print(genes)

## Per sample mode enabled.
plotCounts(object, genes = genes, perSample = TRUE)

## Per sample mode disabled.
plotCounts(object, genes = genes, perSample = FALSE)
}
\seealso{
\itemize{
\item \code{Seurat::DotPlot()}.
\item \code{Seurat::VlnPlot()}.
\item \code{Seurat::RidgePlot()}.
\item \code{monocle3::plot_genes_violin()}.
}
}
