% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMarker-methods.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,SingleCellExperiment-method}
\alias{plotFeature,seurat-method}
\title{Plot Feature}
\usage{
plotFeature(object, ...)

\S4method{plotFeature}{SingleCellExperiment}(object, features, reducedDim,
  color, pointSize, pointAlpha, pointsAsNumbers, label, labelSize, dark,
  legend)

\S4method{plotFeature}{seurat}(object, features, reducedDim, color,
  pointSize, pointAlpha, pointsAsNumbers, label, labelSize, dark, legend)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{legend}{\code{boolean}. Include plot legend.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
data(seurat_small)
plotFeature(seurat_small, features = c("nUMI", "nGene", "mitoRatio"))
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotKnownMarkers}},
  \code{\link{plotMarker}}, \code{\link{plotPCElbow}},
  \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\concept{Plot Functions}
