% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotKnownMarkers-methods.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownSeuratMarkers-method}
\alias{plotKnownMarkers,seurat,KnownSeuratMarkers-method}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(object, markers, ...)


  \S4method{plotKnownMarkers}{SingleCellExperiment,KnownSeuratMarkers}(object,
  markers, reducedDim = "TSNE", headerLevel = 2L, ...)

\S4method{plotKnownMarkers}{seurat,KnownSeuratMarkers}(object, markers,
  reducedDim = "TSNE", headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{grouped_df}. Marker genes, grouped by "\code{cellType}".}

\item{...}{Additional arguments.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Known Markers
}
\examples{
data(seurat_small, known_markers_small)
plotKnownMarkers(
    object = seurat_small,
    markers = known_markers_small
)
}
\seealso{
Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotFeature}},
  \code{\link{plotMarker}}, \code{\link{plotPCElbow}},
  \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\concept{Plot Functions}
