% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/plotKnownMarkersDetected-methods.R
\docType{methods}
\name{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{plotKnownMarkersDetected,SingleCellExperiment-method}
\alias{plotKnownMarkersDetected,seurat-method}
\title{Plot Known Markers Detected}
\usage{
plotKnownMarkersDetected(object, ...)

\S4method{plotKnownMarkersDetected}{SingleCellExperiment}(object, markers,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotKnownMarkersDetected}{seurat}(object, markers,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{markers}{\code{grouped_df}. Marker genes, grouped by "\code{cellType}".}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Known Markers Detected
}
\examples{
plotKnownMarkersDetected(
    object = sce_small,
    markers = known_markers_detected_small
)
}
\seealso{
Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotFeature}},
  \code{\link{plotGene}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}, \code{\link{plotViolin}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
