% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarker-methods.R
\docType{methods}
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarker,SingleCellExperiment-method}
\alias{plotMarker,seurat-method}
\title{Plot cell-type-specific gene markers}
\usage{
\S4method{plotMarker}{SingleCellExperiment}(object, genes,
  reducedDim = "TSNE", expression = c("mean", "sum"),
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_colour_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)

\S4method{plotMarker}{seurat}(object, genes, reducedDim = "TSNE",
  expression = c("mean", "sum"),
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_colour_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes TNSE, UMAP, PCA, for example.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level. Useful when there many points in the dataset
(e.g. single-cell data), and some points can be masked.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}
ggplot2 theme.}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\examples{
data(seurat_small)
object <- seurat_small
title <- "most abundant genes"
genes <- counts(object) \%>\%
    Matrix::rowSums(.) \%>\%
    sort(decreasing = TRUE) \%>\%
    head(n = 4L) \%>\%
    names()
str(genes)

## Default appearance.
plotMarker(object, genes = genes[[1L]])

## Dark mode with viridis palette.
plotMarker(
    object = object,
    genes = genes,
    expression = "mean",
    pointsAsNumbers = TRUE,
    color = ggplot2::scale_color_viridis_c(),
    dark = TRUE,
    label = FALSE,
    title = title
)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
