% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotPcElbow-methods.R
\name{plotPcElbow}
\alias{plotPcElbow}
\alias{plotPcElbow,Seurat-method}
\title{Plot principal component elbow}
\usage{
plotPcElbow(object, ...)

\S4method{plotPcElbow}{Seurat}(object, minPct = 0.01, maxCumPct = 0.9)
}
\arguments{
\item{object}{Object.}

\item{minPct}{\code{numeric(1)} (\code{0}-\code{1}).
Minimum percent standard deviation.}

\item{maxCumPct}{\code{numeric(1)} (\code{0}-\code{1}).
Maximum cumulative percen standard deviation.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
Elbow point is saved in \code{attr(object, "elbow")}.
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\details{
Automatically return the smallest number of PCs that match the \code{minSD},
\code{minPct}, and \code{maxCumPct} cutoffs.
}
\note{
Updated 2023-08-16.
}
\examples{
data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
plotPcElbow(object)
}
\seealso{
\itemize{
\item \code{Seurat::PCElbowPlot()}.
}
}
