% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotReducedDim-methods.R
\docType{methods}
\name{plotReducedDim}
\alias{plotReducedDim}
\alias{plotTSNE}
\alias{plotUMAP}
\alias{plotReducedDim,SingleCellExperiment-method}
\alias{plotReducedDim,seurat-method}
\alias{plotTSNE,SingleCellExperiment-method}
\alias{plotTSNE,seurat-method}
\alias{plotUMAP,SingleCellExperiment-method}
\alias{plotUMAP,seurat-method}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA,seurat-method}
\title{Plot Reduced Dimensions}
\usage{
plotReducedDim(object, ...)

plotTSNE(object, ...)

plotUMAP(object, ...)

\S4method{plotReducedDim}{SingleCellExperiment}(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotReducedDim}{seurat}(object, reducedDim, dimsUse = c(1L,
  2L), interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotTSNE}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotTSNE}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotUMAP}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotUMAP}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotPCA}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotPCA}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = "ident",
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.75),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{dimsUse}{\code{integer}. Vector of length 2 that denotes the columns from
the reduced dimension matrix to use for \code{centerX} and \code{centerY} column
calculations. Defaults the first and second dimensions.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
\itemize{
\item t-SNE: \strong{t}-distributed \strong{S}tochastic \strong{N}eighbor \strong{E}mbedding.
\item PCA: \strong{P}rincipal \strong{C}omponent \strong{A}nalysis.
\item UMAP: \strong{U}niform \strong{M}anifold \strong{A}pproximation and \strong{P}rojection.
}
}
\note{
\code{\link[=plotUMAP]{plotUMAP()}} requires the Python dependency \code{umap-learn}. We recommend
installing this with conda: \code{conda install -c conda-forge umap-learn}.
}
\examples{
object <- sce_small

# t-SNE
plotTSNE(object)
plotTSNE(object, pointsAsNumbers = TRUE, dark = TRUE, label = FALSE)

# UMAP
plotUMAP(object)

# PCA
plotPCA(object)
}
\seealso{
\itemize{
\item \code{\link[Seurat:DimPlot]{Seurat::DimPlot()}}.
\item \href{https://github.com/lmcinnes/umap}{UMAP GitHub repo}.
\item \href{https://satijalab.org/seurat/mca.html}{Seurat Mouse Cell Atlas vignette}.
}

Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotFeature}},
  \code{\link{plotGene}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarker}}, \code{\link{plotPCElbow}},
  \code{\link{plotTopMarkers}}, \code{\link{plotViolin}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Plot Functions}
