% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedBarPlot-methods.R
\name{plotStackedBarPlot}
\alias{plotStackedBarPlot}
\alias{plotCellTypesPerCluster,SingleCellExperiment,ANY-method}
\alias{plotCellTypesPerCluster,Seurat,ANY-method}
\title{Plot stacked bar plot}
\usage{
plotStackedBarPlot(object, ...)

\S4method{plotCellTypesPerCluster}{SingleCellExperiment,ANY}(object)

\S4method{plotCellTypesPerCluster}{Seurat,ANY}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot stacked bar plot
}
\note{
Updated 2020-06-10.
}
\examples{
data(Seurat, package = "acidtest")
data(seurat_known_markers)

## Seurat ====
object <- Seurat
markers <- seurat_known_markers

plotCellTypesPerCluster(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
