% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMarkers-methods.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,Seurat,SeuratMarkersPerCluster-method}
\title{Plot top markers}
\usage{
plotTopMarkers(object, markers, ...)

\S4method{plotTopMarkers}{Seurat,SeuratMarkersPerCluster}(object, markers,
  n = 1L, direction = c("up", "down", "both"), reduction = 1L,
  headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing cell marker expression data.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot top markers
}
\details{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\note{
Updated 2019-07-31.
}
\examples{
data(Seurat, package = "acidtest")
data(seuratAllMarkers)

## Seurat, SeuratMarkersPerCluster ====
object <- Seurat
markers <- seuratAllMarkers
plotTopMarkers(object = object, markers = markers)
}
