% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotViolin-methods.R
\docType{methods}
\name{plotViolin}
\alias{plotViolin}
\alias{plotViolin,SingleCellExperiment-method}
\alias{plotViolin,seurat-method}
\title{Plot Violin}
\usage{
plotViolin(object, ...)

\S4method{plotViolin}{SingleCellExperiment}(object, genes,
  interestingGroups, scale = c("count", "width", "area"),
  color = getOption("pointillism.discrete.color", NULL),
  fill = getOption("pointillism.discrete.fill", NULL),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotViolin}{seurat}(object, genes, interestingGroups,
  scale = c("count", "width", "area"),
  color = getOption("pointillism.discrete.color", NULL),
  fill = getOption("pointillism.discrete.fill", NULL),
  legend = getOption("pointillism.legend", TRUE), title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot Violin
}
\examples{
object <- sce_small
# Plotting with either gene IDs or gene names (symbols) works.
geneIDs <- head(rownames(object), n = 4L)
geneNames <- head(rowRanges(object)[["geneName"]])
glimpse(genes)
plotViolin(object, genes = genes)
}
\seealso{
\code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}.

Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotFeature}},
  \code{\link{plotGene}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarker}}, \code{\link{plotPCElbow}},
  \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Plot Functions}
