% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pointillism-package}
\alias{pointillism}
\alias{pointillism-package}
\title{pointillism}
\description{
R package for single-cell RNA-seq clustering analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/pointillism/}
  \item \url{https://github.com/acidgenomics/r-pointillism/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-pointillism/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})
}

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID}) [contributor]
}

}
\keyword{internal}
