% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk-methods.R
\name{pseudobulk}
\alias{pseudobulk}
\alias{pseudobulk,SingleCellExperiment-method}
\title{Pseudobulk}
\usage{
pseudobulk(object, ...)

\S4method{pseudobulk}{SingleCellExperiment}(object, fun = c("sum", "mean"))
}
\arguments{
\item{object}{Object.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with cells aggregated to samples, like a bulk
RNA-seq dataset.
}
\description{
Pseudobulk
}
\note{
Updated 2020-01-30.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
x <- pseudobulk(object)
print(x)
}
