% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankedList-methods.R
\name{RankedList}
\alias{RankedList}
\alias{RankedList,SeuratMarkersPerCluster-method}
\title{Ranked list of genes}
\usage{
RankedList(object, ...)

\S4method{RankedList}{SeuratMarkersPerCluster}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{RankedList}.
}
\description{
Class containing parameterized ranked gene lists.
}
\note{
Updated 2020-05-20.
}
\section{Gene symbol multi-mapping}{



Multiple gene IDs can map to a gene symbol (e.g. \emph{Homo sapiens} HGNC names).
In this event, we're averaging the stat values using \code{mean()} internally.

}

\examples{
## DESeqAnalysis ====
data(deseq, package = "DESeqAnalysis")
x <- RankedList(deseq)
print(x)

## FGSEAList ====
data(fgsea, package = "acidgsea")
x <- RankedList(fgsea)
print(x)

## matrix ====
data(matrix_lfc, package = "acidtest")
x <- RankedList(matrix_lfc)
print(x)
}
