% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runSeurat-methods.R
\name{runSeurat}
\alias{runSeurat}
\alias{runSeurat,Seurat-method}
\alias{runSeurat,SingleCellExperiment-method}
\title{Run Seurat analysis}
\usage{
runSeurat(object, ...)

\S4method{runSeurat}{Seurat}(
  object,
  regressCellCycle = c("s-g2m-diff", "yes", "no"),
  varsToRegress = c("nCount_RNA", "mitoRatio"),
  dims = "auto",
  resolution = seq(from = 0.2, to = 1.2, by = 0.2),
  workers = "auto"
)

\S4method{runSeurat}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{regressCellCycle}{\code{character(1)}.
\itemize{
\item \code{"s-g2m-diff"}: Calculate the difference between S and G2/M phases and
use that to regress. See \code{CC.Difference} metric in Seurat vignette.
\item \code{"yes"}: Regress out any effects of both S and G2/M phase variable.
Refer to \code{"S.Score"} and \code{"G2M.Score"} metrics in Seurat vignette.
\item \code{"no"}: Don't calculate cell-cycle scoring and don't regress.
}

Refer to the Seurat cell-cycle regression vignette for details.}

\item{varsToRegress}{\code{character} or \code{NULL}.
Unwanted sources of variance to regress. Note that when \code{regressCellCycle}
is not \code{"no"}, then the corresponding cell-cycle variables are added
automatically. Passes to \link[Seurat:ScaleData]{Seurat::ScaleData} internally.}

\item{dims}{\code{"auto"} or \code{integer}.
Dimensions of reduction to use as input for shared nearest neighbor (SNN)
graph construction. When set to "auto" (default), the elbow point is
calculated internally. See \code{\link[=plotPcElbow]{plotPcElbow()}} for details. Passes to
\code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}} and \code{\link[Seurat:RunUMAP]{Seurat::RunUMAP()}} internally.}

\item{resolution}{\code{numeric}.
Resolutions to calculate for clustering.
Passes to \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}} internally.}

\item{workers}{\code{"auto"}, \code{integer(1)}, or \code{NULL}.
Disable parallelization with future by setting to \code{NULL}.}
}
\value{
\code{Seurat}.
}
\description{
Run Seurat analysis
}
\note{
Updated 2023-09-21.
}
\seealso{
\itemize{
\item https://github.com/satijalab/seurat/wiki
\item https://satijalab.org/seurat/articles/essential_commands.html
\item https://satijalab.org/seurat/articles/cell_cycle_vignette.html
\item https://satijalab.org/seurat/articles/future_vignette.html
\item https://satijalab.org/seurat/articles/pbmc3k_tutorial.html
}
}
