% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbwave.R
\name{runZinbwave}
\alias{runZinbwave}
\title{Calculate ZINB-WaVE Weights}
\usage{
runZinbwave(object, BPPARAM = BiocParallel::SerialParam(),
  epsilon = 1e+12, recalculate = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{\link[BiocParallel]{bpparam}} for details.}

\item{recalculate}{\code{logical}. Force recalculation of weights.}

\item{...}{Passthrough arguments to \code{\link[zinbwave:zinbwave]{zinbwave::zinbwave()}}.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Attempt to return stashed zinbwave calcs, or recalculate.
}
\details{
zinbwave will calculate \code{normalizedValues} and \code{weights} matrices, which will
be defined in the \code{\link[=assays]{assays()}} slot.
}
\examples{
library(bcbioSingleCell)
object <- sce_small[seq_len(100L), seq_len(100L)]
object <- filterCells(object)
zinb <- runZinbwave(object)
print(zinb)
assayNames(zinb)
}
\seealso{
\code{\link[zinbwave:zinbwave]{zinbwave::zinbwave()}}.

Other Zero Count Inflation Functions: \code{\link{hasZinbwave}}
}
\author{
Michael Steinbaugh
}
\concept{Zero Count Inflation Functions}
