% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_small}
\alias{sce_small}
\title{SingleCellExperiment Example}
\format{An object of class \code{SingleCellExperiment} with 10000 rows and 100 columns.}
\usage{
sce_small
}
\description{
SingleCellExperiment Example
}
\examples{
sce_small
}
\seealso{
Other Minimal Example Data: \code{\link{all_markers_small}},
  \code{\link{known_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
