% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zzz-seurat-methods.R
\docType{methods}
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{cell2sample}
\alias{seurat-SingleCellExperiment}
\alias{assay,seurat,ANY-method}
\alias{assayNames,seurat-method}
\alias{assays,seurat-method}
\alias{barcodeRanksPerSample,seurat-method}
\alias{cell2sample,seurat-method}
\title{Extend S4 Methods for \code{seurat} Class}
\usage{
barcodeRanksPerSample(object, ...)

cell2sample(object, ...)

\S4method{assay}{seurat,ANY}(x, i, ...)

\S4method{assayNames}{seurat}(x, ...)

\S4method{assays}{seurat}(x, ..., withDimnames = TRUE)

\S4method{barcodeRanksPerSample}{seurat}(object, ...)

\S4method{cell2sample}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{x}{Primary object.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\author{
Michael Steinbaugh
}
\keyword{internal}
