% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat}
\alias{seurat}
\title{Seurat example}
\format{An object of class \code{Seurat} with 230 rows and 80 columns.}
\usage{
seurat
}
\description{
Seurat example
}
\examples{
data(seurat)
seurat
}
\keyword{datasets}
