% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_all_markers}
\alias{seurat_all_markers}
\title{Seurat all markers example}
\format{An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 76 x 37 x 8 x 8 x 8.}
\usage{
seurat_all_markers
}
\description{
Seurat all markers example
}
\examples{
data(seurat_all_markers)
seurat_all_markers
}
\keyword{datasets}
