% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_known_markers}
\alias{seurat_known_markers}
\title{Seurat known markers example}
\format{An object of class \code{KnownMarkers} with 16 rows and 10 columns.}
\usage{
data(seurat_known_markers)
}
\value{
\code{KnownMarkers}.
}
\description{
Seurat known markers example
}
\note{
Updated 2020-01-30.
}
\examples{
data(seurat_known_markers)
seurat_known_markers
}
\keyword{datasets}
