% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_small}
\alias{seurat_small}
\title{Seurat Example}
\format{An object of class \code{seurat} of length 1.}
\usage{
seurat_small
}
\description{
Seurat Example
}
\examples{
data(seurat_small)
seurat_small
}
\keyword{datasets}
