% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratAllMarkers}
\alias{seuratAllMarkers}
\title{Seurat all markers example}
\format{An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 76 x 37 x 8 x 8 x 8.}
\usage{
seuratAllMarkers
}
\description{
Seurat all markers example
}
\examples{
data(seuratAllMarkers)
seuratAllMarkers
}
\keyword{datasets}
