% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratKnownMarkers}
\alias{seuratKnownMarkers}
\title{Seurat known markers example}
\format{An object of class \code{KnownMarkers} with 16 rows and 10 columns.}
\usage{
seuratKnownMarkers
}
\description{
Seurat known markers example
}
\examples{
data(seuratKnownMarkers)
seuratKnownMarkers
}
\keyword{datasets}
