% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/SeuratMarkers-methods.R
\docType{class}
\name{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster}
\alias{SeuratMarkersPerCluster,data.frame-method}
\title{Seurat markers per cluster}
\usage{
SeuratMarkersPerCluster(object, ...)

\S4method{SeuratMarkersPerCluster}{data.frame}(object, ranges,
  alpha = 0.05)
}
\value{
\code{SeuratMarkersPerCluster}.
Results are split per \code{cluster} and arranged by adjusted \emph{P} value
(\code{padj}).
}
\description{
Class containing essential elements for marker per cluster analysis.
}
\note{
Updated 2019-08-06.
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 classes}
