% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SeuratMarkersPerCluster}
\alias{SeuratMarkersPerCluster}
\title{Seurat markers per cluster}
\format{
An object of class \code{nonstandardGenericFunction} of length 1.
}
\usage{
data(SeuratMarkersPerCluster)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Seurat markers per cluster
}
\note{
Updated 2022-03-10.
}
\examples{
data(SeuratMarkersPerCluster)
print(SeuratMarkersPerCluster)
}
\keyword{datasets}
