% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactors-methods.R
\docType{methods}
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{sizeFactors,cell_data_set-method}
\alias{sizeFactors<-,cell_data_set,numeric-method}
\title{Size factors}
\usage{
sizeFactors(object, ...)

sizeFactors(object, ...) <- value

\S4method{sizeFactors}{cell_data_set}(object, type = NULL)

\S4method{sizeFactors}{cell_data_set,numeric}(object,
  type = NULL) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to be assigned to corresponding components of object.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
Named numeric vector, corresponding to cells.
}
\description{
Size factors
}
\note{
Updated 2019-08-05.
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
}
\seealso{
\itemize{
\item \code{monocle3::size_factors()}.
}
}
