% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smpc}
\alias{smpc}
\title{Example \code{SeuratMarkersPerCluster} object}
\format{
An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 79 x 37 x 8 x 8 x 8.
}
\usage{
data(smpc)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Example \code{SeuratMarkersPerCluster} object
}
\note{
Updated 2022-05-11.
}
\examples{
data(smpc)
print(smpc)
}
\keyword{datasets}
