% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,SeuratMarkers-method}
\alias{summary,SeuratMarkersPerCluster-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{SeuratMarkers}(object)

\S4method{summary}{SeuratMarkersPerCluster}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Object summaries
}
\note{
Updated 2019-07-31.
}
\examples{
data(seuratAllMarkers)

## SeuratMarkersPerCluster ====
object <- seuratAllMarkers
class(object)
summary(object)
}
