% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topMarkers-methods.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,SeuratMarkers-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{SeuratMarkers}(object, n = 10L,
  direction = c("positive", "negative", "both"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Number of genes per cluster.}

\item{direction}{\code{string}. Whether to include "\code{positive}", "\code{negative}", or
"\code{both}" directions of association per cluster.}

\item{coding}{\code{boolean}. Include only protein coding genes.}
}
\value{
\code{grouped_df}.
}
\description{
Top Markers
}
\examples{
x <- topMarkers(
    all_markers_small,
    n = 2L,
    direction = "positive"
)
print(x)
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}.
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}

Other Marker Functions: \code{\link{CellTypeMarkers}},
  \code{\link{SeuratMarkers}}, \code{\link{clusterID}},
  \code{\link{knownMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Marker Functions}
