% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_markers_small}
\alias{all_markers_small}
\title{Sanitized Markers Example}
\format{An object of class \code{SeuratMarkers} with 607 rows and 8 columns.}
\usage{
all_markers_small
}
\description{
Sanitized Markers Example
}
\examples{
pointillism::all_markers_small \%>\% glimpse()
}
\seealso{
Other Minimal Example Data: \code{\link{known_markers_small}},
  \code{\link{seurat_small}}
}
\author{
Michael Steinbaugh
}
\concept{Minimal Example Data}
\keyword{datasets}
