% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_markers_small}
\alias{all_markers_small}
\title{Sanitized Markers Example}
\format{An object of class \code{SeuratMarkers} with 208 rows and 8 columns.}
\usage{
all_markers_small
}
\description{
Sanitized Markers Example
}
\examples{
data(all_markers_small)
all_markers_small
}
\keyword{datasets}
