% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/as.SingleCellExperiment-methods.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\alias{as.SingleCellExperiment,Seurat-method}
\title{Coerce object to \code{SingleCellExperiment}}
\usage{
as.SingleCellExperiment(x, ...)

\S4method{as.SingleCellExperiment}{Seurat}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
\code{DESeqDataSet}.
}
\description{
Coerce object to \code{SingleCellExperiment}
}
\details{
S4 coercion support for creating a \code{SingleCellExperiment} from a \code{Seurat}
class object. The \href{https://satijalab.org/Seurat/faq}{Seurat FAQ page}
explains the \code{Seurat} S4 class structure in detail. Internally, this method
improves the basic \code{Seurat::as.SingleCellExperiment} S3 coercion method,
including the \code{object@scale.data} matrix, and will keep track of stashed
\code{rowRanges} and \code{metadata} if the \code{Seurat} object was originally created
from a \code{SingleCellExperiment} (i.e. from the bcbioSingleCell package).
}
\note{
Updated 2022-05-11.
}
\examples{
data(seurat)

## Seurat to SingleCellExperiment ====
from <- seurat
to <- as.SingleCellExperiment(from)
class(to)
print(to)
}
\seealso{
\itemize{
\item \code{Seurat::as.SingleCellExperiment()}.
}
}
