% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-Seurat-methods.R
\name{base-Seurat}
\alias{base-Seurat}
\alias{Gene2Symbol,Seurat-method}
\alias{assay,Seurat,ANY-method}
\alias{assayNames,Seurat-method}
\alias{assays,Seurat-method}
\alias{colData,Seurat-method}
\alias{colData<-,Seurat,DataFrame-method}
\alias{counts,Seurat-method}
\alias{interestingGroups,Seurat-method}
\alias{interestingGroups<-,Seurat,character-method}
\alias{logcounts,Seurat-method}
\alias{mapGenesToIDs,Seurat-method}
\alias{mapGenesToRownames,Seurat-method}
\alias{mapGenesToSymbols,Seurat-method}
\alias{metadata,Seurat-method}
\alias{metadata<-,Seurat-method}
\alias{metrics,Seurat-method}
\alias{normcounts,Seurat-method}
\alias{organism,Seurat-method}
\alias{organism<-,Seurat-method}
\alias{reducedDim,Seurat,ANY-method}
\alias{reducedDimNames,Seurat-method}
\alias{reducedDims,Seurat-method}
\alias{rowData,Seurat-method}
\alias{rowRanges,Seurat-method}
\alias{rowRanges<-,Seurat,ANY-method}
\alias{sampleData,Seurat-method}
\alias{sampleData<-,Seurat,DataFrame-method}
\alias{sampleNames,Seurat-method}
\title{Extend base S4 methods for \code{Seurat} class}
\usage{
\S4method{Gene2Symbol}{Seurat}(object, ...)

\S4method{assay}{Seurat,ANY}(x, i, withDimnames = TRUE, ...)

\S4method{assayNames}{Seurat}(x, ...)

\S4method{assays}{Seurat}(x, withDimnames = TRUE, ...)

\S4method{colData}{Seurat}(x, ...)

\S4method{colData}{Seurat,DataFrame}(x) <- value

\S4method{counts}{Seurat}(object, assay = NULL)

\S4method{interestingGroups}{Seurat}(object, ...)

\S4method{interestingGroups}{Seurat,character}(object) <- value

\S4method{logcounts}{Seurat}(object, assay = NULL)

\S4method{mapGenesToIDs}{Seurat}(object, genes, strict = TRUE)

\S4method{mapGenesToRownames}{Seurat}(object, genes, strict = TRUE)

\S4method{mapGenesToSymbols}{Seurat}(object, genes, strict = TRUE)

\S4method{metadata}{Seurat}(x, ...)

\S4method{metadata}{Seurat}(x) <- value

\S4method{metrics}{Seurat}(object, ...)

\S4method{normcounts}{Seurat}(object, assay = NULL)

\S4method{organism}{Seurat}(object)

\S4method{organism}{Seurat}(object) <- value

\S4method{reducedDim}{Seurat,ANY}(x, type = 1L, withDimnames = TRUE)

\S4method{reducedDimNames}{Seurat}(x)

\S4method{reducedDims}{Seurat}(x, ...)

\S4method{rowData}{Seurat}(x, use.names = TRUE, ...)

\S4method{rowRanges}{Seurat}(x)

\S4method{rowRanges}{Seurat,ANY}(x) <- value

\S4method{sampleData}{Seurat}(object, ...)

\S4method{sampleData}{Seurat,DataFrame}(object) <- value

\S4method{sampleNames}{Seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{x}{Object.}

\item{i}{Indices specifying elements to extract or replace. Indices are \code{numeric} or
\code{character} vectors, empty (\code{missing}), or \code{NULL}.

For more information:\preformatted{help(topic = "Extract", package = "base")
}}

\item{value}{Value to assign.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{genes}{\code{character}.
Gene identifiers.}
}
\value{
Varies, depending on the generic.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\note{
Updated 2021-03-03.
}
\seealso{
\itemize{
\item \code{Seurat::GetAssayData()}.
}
}
\keyword{internal}
