% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellCountsPerCluster-methods.R
\docType{methods}
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,seurat-method}
\title{Cell Counts per Cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object,
  interestingGroups = NULL)

\S4method{cellCountsPerCluster}{seurat}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character}. Groups of interest that
define the samples. If left unset, defaults to \code{sampleName}.}
}
\value{
\code{tbl_df}. Grouped by \code{ident} column and arranged by \code{n}.
}
\description{
Cell Counts per Cluster
}
\examples{
data(seurat_small)
x <- cellCountsPerCluster(seurat_small)
print(x)
}
