% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountsPerCluster-methods.R
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,Seurat-method}
\title{Cell counts per cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object)

\S4method{cellCountsPerCluster}{Seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell counts per cluster
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
x <- cellCountsPerCluster(object)
print(x)
}
