% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountsPerCluster-methods.R
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,Seurat-method}
\alias{cellCountsPerCluster,cell_data_set-method}
\title{Cell counts per cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object)

\S4method{cellCountsPerCluster}{Seurat}(object)

\S4method{cellCountsPerCluster}{cell_data_set}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell counts per cluster
}
\note{
Updated 2019-10-30.
}
\examples{
data(
    Seurat,
    cell_data_set,
    package = "acidtest"
)

## Seurat ====
object <- Seurat
x <- cellCountsPerCluster(object)
print(x)

## cell_data_set ===
## > object <- cell_data_set
## > x <- cellCountsPerCluster(object)
## > print(x)
}
