% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellCountsPerCluster-methods.R
\docType{methods}
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,seurat-method}
\title{Cell Counts per Cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object,
  interestingGroups = NULL)

\S4method{cellCountsPerCluster}{seurat}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
\code{tbl_df}. Grouped by \code{ident} column and arranged by \code{n}.
}
\description{
Cell Counts per Cluster
}
\examples{
x <- cellCountsPerCluster(sce_small)
print(x)
}
\seealso{
Other Cluster Statistics Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{clusterCellCountsPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Cluster Statistics Functions}
