% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCountsPerCluster-methods.R
\docType{methods}
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\alias{cellCountsPerCluster,Seurat-method}
\title{Cell counts per cluster}
\usage{
\S4method{cellCountsPerCluster}{SingleCellExperiment}(object,
  interestingGroups = NULL)

\S4method{cellCountsPerCluster}{Seurat}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
\code{tbl_df}. Grouped by \code{ident} column and arranged by \code{n}.
}
\description{
Cell counts per cluster
}
\examples{
data(seurat)
x <- cellCountsPerCluster(seurat)
print(x)
}
