% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{Cell-Cycle Markers}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.
}
\seealso{
\code{\link[=CellCycleMarkers]{CellCycleMarkers()}}.

Other S4 classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownSeuratMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 classes}
