% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{Cell-cycle markers}
\value{
\code{CellTypeMarkers}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[S4Vectors:Annotated-class]{metadata()}}.
}
\note{
Updated 2019-09-03.
}
