% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/CellMarkers.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\alias{CellCycleMarkers}
\title{Cell-cycle markers}
\usage{
CellCycleMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}
}
\value{
\code{CellTypeMarkers}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[S4Vectors:metadata]{metadata()}}.
}
\seealso{
Other S4 classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
\concept{S4 classes}
