% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{Cell-Cycle Markers}
\value{
\code{CellTypeMarkers}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{metadata()}.
}
\seealso{
\code{CellCycleMarkers()}.

Other S4 classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
\concept{S4 classes}
