% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{\code{CellCycleMarkers} Class}
\description{
\code{CellCycleMarkers} Class
}
\seealso{
Other S4 Object: \code{\link{CellTypeMarkers-class}},
  \code{\link{SingleCellMarkers-class}},
  \code{\link{coerce}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
