% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellCycleMarkers-class}
\alias{CellCycleMarkers-class}
\title{\code{CellCycleMarkers} Class}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=metadata]{metadata()}}.
}
\seealso{
\code{\link[=CellCycleMarkers]{CellCycleMarkers()}}.

Other S4 Classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownSeuratMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 Classes}
