% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\title{Cell-cycle markers}
\usage{
CellCycleMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}
}
\description{
Cell-cycle markers
}
\note{
Updated 2019-07-31.
}
