% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\title{\code{CellCycleMarkers} Class}
\usage{
CellCycleMarkers(object, gene2symbol)
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=metadata]{metadata()}}.
}
\seealso{
\code{\link[=CellCycleMarkers]{CellCycleMarkers()}}.

Other S4 classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownSeuratMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
