% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\title{\code{CellCycleMarkers} Generator}
\usage{
CellCycleMarkers(file, gs, ws = 1L, organism, ensemblRelease,
  gene2symbol = NULL)
}
\arguments{
\item{file}{\code{string} or \code{missing}. Gene markers file path (CSV or Excel).}

\item{gs}{\code{string} or \code{missing}. Google Sheets \code{sheet_key} identifier, which
can be located with the \code{\link[googlesheets:gs_ls]{googlesheets::gs_ls()}} function.}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{organism}{\code{string}. Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{scalar integer}. Ensembl release version (e.g. \code{90}).}

\item{gene2symbol}{\code{gene2symbol} or \code{NULL}. Use the stashed gene-to-symbol
mappings from a \code{SingleCellExperiment} or \code{seurat} object.}
}
\value{
\code{CellCycleMarkers}.
}
\description{
Currently designed for internal use by the pointillism package.
}
\examples{
# Google Sheets method.
x <- CellCycleMarkers(
    gs = "1qA5ktYeimNGpZF1UPSQZATbpzEqgyxN6daoMOjv6YYw",
    ws = "Homo_sapiens",
    organism = "Homo sapiens",
    ensemblRelease = 92L
)
}
\author{
Michael Steinbaugh
}
\concept{S4 Generators}
