% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCycleMarkers.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\alias{importCellCycleMarkers}
\title{Cell-cycle markers}
\usage{
CellCycleMarkers(object, gene2symbol)

importCellCycleMarkers(file, organism, release)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{file}{\code{character(1)}.
File path.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{CellCycleMarkers}.
}
\description{
Cell-cycle markers
}
\note{
Updated 2020-02-21.
}
\examples{
markers_dir <- system.file(
    file.path("extdata", "markers"),
    package = "pointillism"
)

cell_cycle_dir <- file.path(markers_dir, "cell-cycle")
files <- list.files(cell_cycle_dir, pattern = "*.csv", full.names = TRUE)
file <- files[[1L]]

## Ensembl release version.
release_file <- file.path(markers_dir, "ensembl-release.txt")
release <- as.integer(readLines(release_file))

importCellCycleMarkers(
    file = file,
    organism = basenameSansExt(file),
    release = release
)
}
