% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCycleMarkers.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\alias{importCellCycleMarkers}
\title{Cell-cycle markers}
\usage{
CellCycleMarkers(object, gene2symbol)

importCellCycleMarkers(file, organism, release, ignoreVersion = TRUE)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneId} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}

\item{file}{\code{character(1)}.
File path.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}

\item{ignoreVersion}{\code{logical(1)}.
Ignore identifier (e.g. transcript, gene) versions.
When applicable, the identifier containing version numbers will be stored
in \code{txIdVersion} and \code{geneIdVersion}, and the variants without versions
will be stored in \code{txId}, \code{txIdNoVersion}, \code{geneId}, and \code{geneIdNoVersion}.}
}
\value{
\code{CellCycleMarkers}.
}
\description{
Cell-cycle markers
}
\note{
Updated 2021-03-02.
}
\examples{
markersDir <- system.file(
    file.path("extdata", "markers"),
    package = "pointillism"
)

cellCycleDir <- file.path(markersDir, "cell-cycle")
files <- list.files(cellCycleDir, pattern = "*.csv", full.names = TRUE)
file <- files[[1L]]

organism <- sentenceCase(gsub("-", " ", basenameSansExt(file)))

## Ensembl release version.
releaseFile <- file.path(markersDir, "ensembl-release.txt")
release <- as.integer(readLines(releaseFile))

importCellCycleMarkers(
    file = file,
    organism = organism,
    release = release
)
}
