% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellCycleMarkers.R
\name{CellCycleMarkers}
\alias{CellCycleMarkers}
\title{Cell-Cycle Markers}
\usage{
CellCycleMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings.}
}
\value{
\code{CellTypeMarkers}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[S4Vectors:metadata]{S4Vectors::metadata()}}.
}
\seealso{
\code{\link[=CellCycleMarkers]{CellCycleMarkers()}}.

Other S4 classes: \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
