% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellCycleMarkersList}
\alias{cellCycleMarkersList}
\title{Cell-cycle markers list}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(cellCycleMarkersList)
}
\value{
\code{list}.
}
\description{
Cell-cycle markers list
}
\note{
Updated 2021-03-03.
}
\examples{
data(cellCycleMarkersList)
names(cellCycleMarkersList)
}
\keyword{datasets}
