% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellTypeMarkers-class}
\alias{CellTypeMarkers-class}
\title{Cell-Type Markers}
\value{
\code{CellTypeMarkers}
}
\description{
Data provenence information, including the organism and Ensembl release are
defined in \code{\link[=metadata]{metadata()}}.
}
\seealso{
\code{\link[=CellTypeMarkers]{CellTypeMarkers()}}.

Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
\concept{S4 classes}
