% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellTypeMarkers-class}
\alias{CellTypeMarkers-class}
\title{\code{CellTypeMarkers} Class}
\description{
\code{CellTypeMarkers} Class
}
\seealso{
Other S4 Object: \code{\link{CellCycleMarkers-class}},
  \code{\link{SingleCellMarkers-class}},
  \code{\link{coerce}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
