% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellTypeMarkers}
\alias{CellTypeMarkers}
\title{Cell-type markers}
\usage{
CellTypeMarkers(object, gene2symbol)
}
\arguments{
\item{object}{Object.}

\item{gene2symbol}{\code{Gene2Symbol}.
Gene-to-symbol mappings. Must contain \code{geneID} and \code{geneName} columns. See
\code{Gene2Symbol} for more information.}
}
\description{
Cell-type markers
}
\note{
Updated 2019-07-31.
}
