% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeMarkers.R
\name{CellTypeMarkers}
\alias{CellTypeMarkers}
\title{Cell-type markers}
\usage{
CellTypeMarkers(file, organism, release)
}
\arguments{
\item{file}{\code{character(1)}.
File path.}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{release}{\code{integer(1)}.
Ensembl release version (e.g. \code{90}). We recommend setting this value if
possible, for improved reproducibility. When left unset, the latest release
available via AnnotationHub/ensembldb is used. Note that the latest version
available can vary, depending on the versions of AnnotationHub and
ensembldb in use.}
}
\value{
\code{CellTypeMarkers}.
}
\description{
Cell-type markers
}
\note{
Updated 2020-02-20.
}
