% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellTypeMarkers}
\alias{CellTypeMarkers}
\title{Cell-Type Markers}
\usage{
CellTypeMarkers(file, gs, ws = 1L, organism, ensemblRelease,
  gene2symbol = NULL)
}
\arguments{
\item{file}{\code{string} or \code{missing}. Gene markers file path (CSV or Excel).}

\item{gs}{\code{string} or \code{missing}. Google Sheets \code{sheet_key} identifier, which
can be located with the \code{\link[googlesheets:gs_ls]{googlesheets::gs_ls()}} function.}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{organism}{\code{string}. Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{scalar integer}. Ensembl release version (e.g. \code{90}).}

\item{gene2symbol}{\code{gene2symbol} or \code{NULL}. Use the stashed gene-to-symbol
mappings from a \code{SingleCellExperiment} or \code{seurat} object.}
}
\value{
\code{CellTypeMarkers}.
}
\description{
Local spreadsheets (CSV, Excel) and Google Sheets are currently supported.
}
\details{
Must contain \code{phase} and \code{geneID} columns.
}
\examples{
# Google Sheets method.
x <- CellTypeMarkers(
    gs = "1vGNU2CCxpaoTCLvzOxK1hf5gjULrf2-CpgCp9bOfGJ0",
    ws = "Homo_sapiens",
    organism = "Homo sapiens",
    ensemblRelease = 92L
)
}
\seealso{
\itemize{
\item \code{\link[googlesheets:gs_key]{googlesheets::gs_key()}}.
\item \code{\link[googlesheets:gs_read]{googlesheets::gs_read()}}.
}

Other Marker Functions: \code{\link{SeuratMarkers}},
  \code{\link{clusterID}}, \code{\link{knownMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Marker Functions}
