% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesPerCluster.R
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(data, min = 1L, max = Inf)
}
\arguments{
\item{min}{\code{scalar integer}. Minimum number of marker genes per cluster.}

\item{max}{\code{scalar integer}. Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df} grouped by "\code{cluster}", containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
x <- cellTypesPerCluster(known_markers_small)
glimpse(x)
}
\seealso{
Other Clustering Functions: \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotMarker}}, \code{\link{plotPCElbow}},
  \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
