% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypesPerCluster-methods.R
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,KnownMarkers-method}
\title{Cell types per cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{KnownMarkers}(object, min = 1L, max = Inf)
}
\arguments{
\item{object}{Object.}

\item{min}{\code{integer(1)}.
Minimum number of marker genes per cluster.}

\item{max}{\code{integer(1)}.
Maximum number of marker genes per cluster.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell types per cluster
}
\note{
Updated 2020-01-30.
}
\examples{
data(cell_type_markers_list, seurat_all_markers)

## KnownMarkers ====
markers <- KnownMarkers(
    markers = seurat_all_markers,
    known = cell_type_markers_list[["homoSapiens"]]
)
x <- cellTypesPerCluster(markers)
print(x)
}
