% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypesPerCluster-methods.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,KnownSeuratMarkers-method}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{KnownSeuratMarkers}(object, min = 1L,
  max = Inf)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{min}{\code{scalar integer}. Minimum number of marker genes per cluster.}

\item{max}{\code{scalar integer}. Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df}. Grouped by \code{cluster} column, containing the count (\code{n})
of significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
x <- cellTypesPerCluster(known_markers_small)
print(x)
}
\seealso{
Other Cluster Statistics Functions: \code{\link{cellCountsPerCluster}},
  \code{\link{clusterCellCountsPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Cluster Statistics Functions}
