% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesPerCluster.R
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\title{Cell Types per Cluster}
\usage{
cellTypesPerCluster(data, min = 1L, max = Inf)
}
\arguments{
\item{data}{\code{data.frame} grouped by \code{cellType}. \code{\link[=knownMarkersDetected]{knownMarkersDetected()}}
return.}

\item{min}{\code{scalar integer}. Minimum number of marker genes per cluster.}

\item{max}{\code{scalar integer}. Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df} grouped by "\code{cluster}", containing the count (\code{n}) of
significant known makers per cell type.
}
\description{
Cell Types per Cluster
}
\examples{
x <- cellTypesPerCluster(known_markers_detected_small)
glimpse(x)
}
\seealso{
Other Cluster Statistics Functions: \code{\link{cellCountsPerCluster}},
  \code{\link{clusterCellCountsPerSample}}
}
\author{
Michael Steinbaugh
}
\concept{Cluster Statistics Functions}
