% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesPerCluster-methods.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,KnownMarkers-method}
\title{Cell Types per Cluster}
\usage{
\S4method{cellTypesPerCluster}{KnownMarkers}(object, min = 1L,
  max = Inf)
}
\arguments{
\item{object}{Object.}

\item{min}{\code{integer(1)}.
Minimum number of marker genes per cluster.}

\item{max}{\code{integer(1)}.
Maximum number of marker genes per cluster.}
}
\value{
\code{grouped_df}.
Grouped by \code{cluster} column, containing the count (\code{n}) of significant  known
makers per cell type.
}
\description{
Cell types per cluster
}
\examples{
data(all_markers_small, cell_type_markers)
markers <- KnownMarkers(
    markers = all_markers_small,
    known = cell_type_markers$homoSapiens
)
x <- cellTypesPerCluster(markers)
print(x)
}
