% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypesPerCluster-methods.R
\docType{methods}
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,KnownMarkers-method}
\title{Cell types per cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{KnownMarkers}(object, min = 1L,
  max = Inf)
}
\arguments{
\item{object}{Object.}

\item{min}{\code{integer(1)}.
Minimum number of marker genes per cluster.}

\item{max}{\code{integer(1)}.
Maximum number of marker genes per cluster.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
Grouped by \code{cluster} column, containing the count (\code{n}) of significant  known
makers per cell type.
}
\description{
Cell types per cluster
}
\note{
Updated 2019-07-31.
}
\examples{
data(cellTypeMarkersList, seuratAllMarkers)
markers <- KnownMarkers(
    markers = seuratAllMarkers,
    known = cellTypeMarkersList[["homoSapiens"]]
)
x <- cellTypesPerCluster(markers)
print(x)
}
