% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCellCountsPerSample-methods.R
\docType{methods}
\name{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample,SingleCellExperiment-method}
\alias{clusterCellCountsPerSample,Seurat-method}
\title{Cluster cell counts per sample}
\usage{
clusterCellCountsPerSample(object, ...)

\S4method{clusterCellCountsPerSample}{SingleCellExperiment}(object)

\S4method{clusterCellCountsPerSample}{Seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{grouped_df}.
Grouped by \code{sampleName} column, arranged by abundance.
}
\description{
Cluster cell counts per sample
}
\note{
Updated 2019-07-31.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
x <- clusterCellCountsPerSample(object)
print(x)
}
