% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCellCountsPerSample-methods.R
\docType{methods}
\name{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample,SingleCellExperiment-method}
\alias{clusterCellCountsPerSample,Seurat-method}
\alias{clusterCellCountsPerSample,cell_data_set-method}
\title{Cluster cell counts per sample}
\usage{
clusterCellCountsPerSample(object, ...)

\S4method{clusterCellCountsPerSample}{SingleCellExperiment}(object)

\S4method{clusterCellCountsPerSample}{Seurat}(object)

\S4method{clusterCellCountsPerSample}{cell_data_set}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cluster cell counts per sample
}
\note{
Updated 2019-09-03.
}
\examples{
data(
    Seurat,
    cell_data_set,
    package = "acidtest"
)

## Seurat ====
object <- Seurat
x <- clusterCellCountsPerSample(object)
print(x)

## cell_data_set ====
object <- cell_data_set
x <- clusterCellCountsPerSample(object)
print(x)
}
