% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/clusterCellCountsPerSample-methods.R, R/ident-methods.R
\docType{methods}
\name{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample}
\alias{clusterCellCountsPerSample,SingleCellExperiment-method}
\alias{clusterCellCountsPerSample,seurat-method}
\alias{ident,seurat-method}
\title{Cluster Cell Counts per Sample}
\usage{
clusterCellCountsPerSample(object, ...)

\S4method{clusterCellCountsPerSample}{SingleCellExperiment}(object)

\S4method{clusterCellCountsPerSample}{seurat}(object)

\S4method{ident}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{tibble} grouped by "\code{sampleName}" column, arranged by abundance.
}
\description{
Cluster Cell Counts per Sample
}
\examples{
clusterCellCountsPerSample(sce_small)
}
