% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clusterID-methods.R
\docType{methods}
\name{clusterID}
\alias{clusterID}
\alias{clusterID,SingleCellExperiment-method}
\alias{clusterID,seurat-method}
\title{Cluster Identity}
\usage{
clusterID(object, ...)

\S4method{clusterID}{SingleCellExperiment}(object)

\S4method{clusterID}{seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster Identity
}
\examples{
clusterID(sce_small)
}
\author{
Michael Steinbaugh
}
\concept{Marker Functions}
