% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterID-methods.R
\docType{methods}
\name{clusterID}
\alias{clusterID}
\alias{clusterID,SingleCellExperiment-method}
\alias{clusterID,Seurat-method}
\title{Cluster identifier}
\usage{
\S4method{clusterID}{SingleCellExperiment}(object)

\S4method{clusterID}{Seurat}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{factor}.
}
\description{
Cluster identifier
}
\examples{
data(seurat_small)
x <- clusterID(seurat_small)
head(x)
table(x)
}
