% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterID-methods.R
\docType{methods}
\name{clusterID}
\alias{clusterID}
\alias{clusterID,SingleCellExperiment-method}
\alias{clusterID,Seurat-method}
\title{Cluster identifier}
\usage{
clusterID(object, ...)

\S4method{clusterID}{SingleCellExperiment}(object)

\S4method{clusterID}{Seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster identifier
}
\note{
Updated 2019-07-31.
}
\examples{
data(seurat)
x <- clusterID(seurat)
head(x)
table(x)
}
