% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters-methods.R
\docType{methods}
\name{clusters}
\alias{clusters}
\alias{clusters,Seurat-method}
\alias{clusters,cell_data_set-method}
\alias{clusters,SingleCellExperiment-method}
\title{Cluster identifiers}
\usage{
clusters(object, ...)

\S4method{clusters}{Seurat}(object)

\S4method{clusters}{cell_data_set}(object, reductionMethod = "UMAP")

\S4method{clusters}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster identifiers
}
\note{
Updated 2019-08-02.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
x <- clusters(object)
head(x)
table(x)
}
