% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/clusters-methods.R
\name{clusters}
\alias{clusters}
\alias{clusters,SingleCellExperiment-method}
\alias{clusters,Seurat-method}
\title{Cluster identifiers}
\usage{
clusters(object, ...)

\S4method{clusters}{SingleCellExperiment}(object)

\S4method{clusters}{Seurat}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{factor}.
}
\description{
Cluster identifiers
}
\note{
Updated 2019-08-02.
}
\examples{
data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
x <- clusters(object)
head(x)
table(x)
}
\seealso{
\itemize{
\item \code{Seurat::Idents()}.
\item \code{monocle3::clusters()}.
}
}
