% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce}
\alias{coerce}
\alias{as}
\alias{coerce,SingleCellExperiment,seurat-method}
\alias{coerce,seurat,SingleCellExperiment-method}
\alias{coerce,seurat,RangedSummarizedExperiment-method}
\alias{coerce,seurat,SummarizedExperiment-method}
\title{Methods for Coercing an Object to a Class}
\value{
Object of new class.
}
\description{
Methods for Coercing an Object to a Class
}
\section{\code{SingleCellExperiment} to \code{seurat}}{

Interally \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}} is called without applying any
additional filtering cutoffs, since we have already defined them during our
quality control analysis. Here we are passing the raw gene-level counts of
the filtered cells into a new \code{seurat} class object. Use
\code{\link[=convertGenesToSymbols]{convertGenesToSymbols()}} to convert gene IDs to names (symbols).
}

\section{\code{seurat} to \code{SingleCellExperiment}}{

S4 coercion support for creating a \code{SingleCellExperiment} from a \code{seurat}
class object. The \href{https://satijalab.org/seurat/faq}{Seurat FAQ page}
explains the \code{seurat} S4 class structure in detail. Internally, this method
improves the basic \code{Seurat::as.SingleCellExperiment()} S3 coercion method,
including the \code{object@scale.data} matrix, and will keep track of stashed
\code{rowRanges} and \code{metadata} if the \code{seurat} object was originally created
from a \code{SingleCellExperiment} (i.e. from the bcbioSingleCell package).
}

\section{\code{seurat} to \code{RangedSummarizedExperiment}}{

S4 coercion support for creating a \code{RangedSummarizedExperiment} from a
\code{seurat} class object.
}

\section{\code{seurat} to \code{SummarizedExperiment}}{

S4 coercion support for creating a \code{SummarizedExperiment} from a \code{seurat}
class object.
}

\examples{
# SingleCellExperiment to seurat ====
x <- as(sce_small, "seurat")
class(x)
print(x)
# seurat to SingleCellExperiment ====
x <- as(seurat_small, "SingleCellExperiment")
print(x)
# seurat to RangedSummarizedExperiment ====
x <- as(seurat_small, "RangedSummarizedExperiment")
print(x)
# seurat to SummarizedExperiment ====
x <- as(seurat_small, "SummarizedExperiment")
print(x)
}
\seealso{
\itemize{
\item \code{\link[methods:as]{methods::as()}}.
\item \code{\link[methods:canCoerce]{methods::canCoerce()}}.
\item \code{\link[Seurat:CreateSeuratObject]{Seurat::CreateSeuratObject()}}.
\item \code{\link[Seurat:Convert]{Seurat::Convert()}}.
}

Other S4 Object: \code{\link{show}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
