% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-cpm-methods.R
\docType{methods}
\name{cpm}
\alias{cpm}
\alias{cpm,SingleCellExperiment-method}
\alias{cpm,Seurat-method}
\title{Counts per million}
\usage{
cpm(object, ...)

\S4method{cpm}{SingleCellExperiment}(object)

\S4method{cpm}{Seurat}(object, assay = NULL)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}
}
\description{
Counts per million
}
\examples{
data(
    SingleCellExperiment,
    Seurat,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
x <- cpm(object)
class(x)
mean(x)

## Seurat ====
object <- Seurat
cpm(object)
}
\seealso{
\itemize{
\item \code{SingleCellExperiment::cpm()}.
\item \code{edgeR::cpm()}.
\item \code{scater::calculateCPM()}.
}
}
