% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm-methods.R
\name{cpm}
\alias{cpm}
\alias{cpm,SingleCellExperiment-method}
\alias{cpm,Seurat-method}
\title{Counts per million}
\usage{
cpm(object, ...)

\S4method{cpm}{SingleCellExperiment}(object)

\S4method{cpm}{Seurat}(object, assay = NULL)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}
}
\value{
\code{matrix}
}
\description{
Counts per million
}
\note{
Updated 2020-01-30.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
cpm <- cpm(object)
class(cpm)
mean(cpm)

## Seurat ====
object <- Seurat
cpm <- cpm(object)
class(cpm)
mean(cpm)
}
\seealso{
\itemize{
\item \code{\link[SingleCellExperiment:cpm]{SingleCellExperiment::cpm()}}.
\item \code{\link[edgeR:cpm]{edgeR::cpm()}}.
\item \code{\link[scater:calculateCPM]{scater::calculateCPM()}}.
}
}
\keyword{internal}
