% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm-methods.R
\docType{methods}
\name{cpm}
\alias{cpm}
\alias{cpm,SingleCellExperiment-method}
\alias{cpm,Seurat-method}
\title{Counts per million}
\usage{
cpm(object, ...)

\S4method{cpm}{SingleCellExperiment}(object, verbose = FALSE)

\S4method{cpm}{Seurat}(object, assay = NULL, verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}
}
\value{
\code{matrix}
}
\description{
Counts per million
}
\note{
Updated 2019-10-30.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
cpm <- cpm(object)
class(cpm)
mean(cpm)

## Seurat ====
object <- Seurat
cpm <- cpm(object)
class(cpm)
mean(cpm)

## cell_data_set ====
object <- cell_data_set
cpm <- cpm(object)
class(cpm)
mean(cpm)
}
\seealso{
\itemize{
\item \code{\link[SingleCellExperiment:cpm]{SingleCellExperiment::cpm()}}.
\item \code{\link[edgeR:cpm]{edgeR::cpm()}}.
\item \code{\link[scater:calculateCPM]{scater::calculateCPM()}}.
}
}
\keyword{internal}
