% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{plotKnownMarkers}
\alias{plotTSNEExpressionData}
\alias{readMarkers}
\alias{readMarkersFile}
\title{Defunct Functions}
\usage{
plotKnownMarkers(...)

plotTSNEExpressionData(...)

readMarkers(...)

readMarkersFile(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
