% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{knownMarkersDetected}
\title{Defunct Functions}
\usage{
knownMarkersDetected(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Defunct]{.Defunct()}}.
}
\description{
Defunct Functions
}
\keyword{internal}
