% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{defunct}
\alias{defunct}
\alias{knownMarkers}
\alias{knownMarkersDetected}
\alias{plotKnownMarkersDetected}
\alias{readCellTypeMarkers}
\alias{sanitizeMarkers}
\alias{importCellCycleMarkersFromGoogle}
\alias{importCellTypeMarkersFromGoogle}
\alias{runZinbwave}
\title{Defunct functions}
\usage{
knownMarkers(...)

knownMarkersDetected(...)

plotKnownMarkersDetected(...)

readCellTypeMarkers(...)

sanitizeMarkers(...)

importCellCycleMarkersFromGoogle(...)

importCellTypeMarkersFromGoogle(...)

runZinbwave(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{.Defunct}.
}
\description{
Defunct functions
}
\examples{
a <- function(...) {
    .Defunct("b")
}

withCallingHandlers(
    expr = tryCatch(
        expr = a(1L),
        error = function(e) {
            print(e)
            invisible()
        }
    )
)
}
\seealso{
\link[base:Defunct]{Defunct}.
}
\keyword{internal}
