% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotKnownMarkersDetected}
\alias{readCellTypeMarkers}
\alias{sanitizeSeuratMarkers}
\title{Deprecated Functions}
\usage{
plotKnownMarkersDetected(...)

readCellTypeMarkers(...)

sanitizeSeuratMarkers(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
\code{\link[=.Deprecated]{.Deprecated()}}.
}
\description{
Deprecated Functions
}
\keyword{internal}
