% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExp-methods.R
\docType{methods}
\name{diffExp}
\alias{diffExp}
\alias{diffExp,SingleCellExperiment-method}
\alias{diffExp,Seurat-method}
\title{Differential expression}
\usage{
\S4method{diffExp}{SingleCellExperiment}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, BPPARAM = BiocParallel::SerialParam())

\S4method{diffExp}{Seurat}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{object}{Object.}

\item{numerator}{\code{character}.
Cells to use in the numerator of the contrast (e.g. treatment).}

\item{denominator}{\code{character}.
Cells to use in the denominator of the contrast (e.g. control).}

\item{caller}{\code{character(1)}.
Package to use for differential expression calling. Defaults to \code{"edgeR"}
(faster for large datasets) but \code{"DESeq2"} is also supported.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells required to perform the differential expression
analysis.}

\item{minCellsPerGene}{\code{integer(1)}.
The minimum number of cells where a gene is expressed, to pass low
expression filtering.}

\item{minCountsPerCell}{\code{integer(1)}.
Minimum number of counts per cell for a gene to pass low expression
filtering. The number of cells is defined by \code{minCellsPerGene}.}

\item{BPPARAM}{\code{bpparamClass}.
Back-end method to be used for computations.
See \code{\link[BiocParallel:bpparam]{bpparam}} for details.
Currently only used by DESeq2 but not edgeR for calculations here.}
}
\value{
Varies depending on the \code{caller} argument:
\itemize{
\item \code{caller = "edgeR"}: \code{DEGLRT}.
\item \code{caller = "DESeq2"}: Unshrunken \code{DESeqResults}.
Apply \code{\link[DESeq2:lfcShrink]{DESeq2::lfcShrink()}} if shrunken results are desired.
}
}
\description{
Differential expression
}
\details{
Perform pairwise differential expression across groups of cells. Currently
supports edgeR and DESeq2 as DE callers.
}
\note{
We are currently recommending the ZINB-WaVE method over zingeR, since
it is faster, and has been show to be more sensitive for most single-cell
RNA-seq datasets.
}
\section{DESeq2}{


We're providing preliminary support for DESeq2 as the differential expression
caller. It is currently considerably slower for large datasets than edgeR.

We're trying to follow the conventions used in DESeq2 for contrasts, defining
the name of the factor in the design formula, numerator, and denominator
level for the fold change calculations. See \code{\link[DESeq2:results]{DESeq2::results()}} for details.

Van de Berge and Perraudeau and others have shown the LRT may perform better
for null hypothesis testing, so we use the LRT. In order to use the Wald
test, it is recommended to set \code{useT = TRUE} (\emph{not currently in use}).

For UMI data, for which the expected counts may be very low, the likelihood
ratio test implemented in \code{nbinomLRT()} should be used.

Note that DESeq2 supports \code{weights} in assays automatically, if slotted using
zinbwave, which is no longer recommended for droplet scRNA-seq.
}

\section{edgeR}{


The LRT has been shown to perform better for null hypothesis testing with
droplet scRNA-seq data. Here we are using \code{\link[edgeR:glmLRT]{edgeR::glmLRT()}} internally.

edgeR is currently significantly faster than DESeq2 for large datasets.
}

\section{Seurat conventions}{


Note that Seurat currently uses the convention \code{cells.1} for the numerator
and \code{cells.2} for the denominator. See \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} for details.
}

\examples{
data(seurat)
object <- seurat

## Compare expression in cluster 3 relative to 2.
ident <- clusterID(object)
numerator <- names(ident)[ident == "3"]
summary(numerator)
denominator <- names(ident)[ident == "2"]
summary(denominator)

## edgeR.
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR"
)
class(x)
summary(x)

## DESeq2.
# This will warn about weights with the minimal example.
# x <- diffExp(
#     object = object,
#     numerator = numerator,
#     denominator = denominator,
#     caller = "DESeq2"
# )
# class(x)
# summary(x)
}
\seealso{
\code{\link[Seurat:WhichCells]{Seurat::WhichCells()}}.
}
