% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffExp-methods.R
\docType{methods}
\name{diffExp}
\alias{diffExp}
\alias{diffExp,SingleCellExperiment-method}
\alias{diffExp,seurat-method}
\title{Differential Expression}
\usage{
diffExp(object, ...)

\S4method{diffExp}{SingleCellExperiment}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)

\S4method{diffExp}{seurat}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{numerator}{\code{character}. Cells to use in the numerator of the contrast
(e.g. treatment).}

\item{denominator}{\code{character}. Cells to use in the denominator of the
contrast (e.g. control).}

\item{caller}{\code{string}. Package to use for differential expression calling.
Defaults to \code{"edgeR"} (faster for large datasets) but \code{"DESeq2"} is also
supported.}

\item{minCellsPerGene}{\code{scalar integer}. The minimum number of cells where a
gene is expressed, to pass low expression filtering. Set to \code{0} to disable
(\emph{not recommended}).}

\item{minCountsPerCell}{\code{scalar integer}. Minimum number of counts per cell
for a gene to pass low expression filtering. The number of cells is defined
by \code{minCellsPerGene}. Set to \code{0} to disable (\emph{not recommended}).}
}
\value{
Varies depending on the \code{caller} argument:
\itemize{
\item \code{caller = "edgeR"}: \code{DEGLRT}.
\item \code{caller = "DESeq2"}: Unshrunken \code{DESeqResults}. Use \code{lfcShrink()} if
shrunken results are desired.
}
}
\description{
Perform pairwise differential expression across groups of cells by fitting
to a zero-inflated negative binomial (ZINB) model using the zinbwave package.
Currently supports edgeR and DESeq2 as DE callers.
}
\details{
This step will run a lot faster if you pre-calculate the ZINB weights using
the \code{\link[=runZinbwave]{runZinbwave()}} function, which will stash the weights into the
\code{\link[=assays]{assays()}} slot of the object. Running zinbwave across the entire set of
filtered cells also has greater sensitivity for weight calculations.
}
\note{
We are currently recommending the ZINB-WaVE method over zingeR, since
it is faster, and has been show to be more sensitive for most single-cell
RNA-seq datasets.
}
\section{zinbwave}{

We are currently using an epsilon setting of \code{1e12}, as recommended by the
ZINB-WaVE integration paper. For more information on the zinbwave package,
refer to these materials:
\itemize{
\item \href{https://doi.org/10.1186/s13059-018-1406-4}{zinbwave paper}.
\item \href{https://bit.ly/2wtDdpS}{zinbwave vignette}.
\item \href{https://github.com/mikelove/zinbwave-deseq2}{zinbwave-DESeq2 workflow}.
}
}

\section{edgeR}{

After estimation of the dispersions and posterior probabilities, the
\code{glmWeightedF()} function is used for statistical inference. This is an
adapted function from the \code{glmLRT()} function of edgeR. It uses an F-test for
which the denominator degrees of freedom are by default adjusted according to
the downweighting of excess zeros (\code{ZI = TRUE}). Also, independent filtering
can be performed on the obtained p-values (\code{independentFiltering = TRUE}). We
use the independent filtering strategy that was originally implemented in
DESeq2. By default, the average fitted values are used as a filter criterion.
}

\section{DESeq2}{

We're providing preliminary support for DESeq2 as the differential expression
caller. It is currently considerably slower for large datasets than edgeR.

We're trying to follow the conventions used in DESeq2 for contrasts, defining
the name of the factor in the design formula, numerator, and denominator
level for the fold change calculations. See \code{\link[DESeq2:results]{DESeq2::results()}} for more
information.
}

\section{Seurat conventions}{

Note that Seurat currently uses the convention \code{cells.1} for the numerator
and \code{cells.2} for the denominator. See \code{\link[Seurat:DiffExpTest]{Seurat::DiffExpTest()}} for
additional information.
}

\examples{
data(seurat_small)
object <- as(seurat_small, "SingleCellExperiment")
# Subset example for speed.
object <- [seq_len(100L), seq_len(100L)]
# Ensure genes with all zero counts are filtered.
object <- bcbioSingleCell::filterCells(object)

numerator <- colnames(object)[object$group == "group2"]
str(numerator)
denominator <- colnames(object)[object$group == "group1"]
str(denominator)

# edgeR
x <- suppressMessages(diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR"
))
class(x)
summary(x)

# DESeq2
x <- suppressMessages(diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "DESeq2"
))
class(x)
summary(x)
}
\seealso{
\code{\link[Seurat:WhichCells]{Seurat::WhichCells()}}.

Other Differential Expression Functions: \code{\link{diffExpPerCluster}},
  \code{\link{findMarkers}}, \code{\link{runZinbwave}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
