% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExp-methods.R
\docType{methods}
\name{diffExp}
\alias{diffExp}
\alias{diffExp,SingleCellExperiment-method}
\alias{diffExp,seurat-method}
\title{Differential expression}
\usage{
\S4method{diffExp}{SingleCellExperiment}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, bpparam = BiocParallel::SerialParam())

\S4method{diffExp}{seurat}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, bpparam = BiocParallel::SerialParam())
}
\arguments{
\item{object}{Object.}

\item{numerator}{\code{character}.
Cells to use in the numerator of the contrast (e.g. treatment).}

\item{denominator}{\code{character}.
Cells to use in the denominator of the contrast (e.g. control).}

\item{caller}{\code{character(1)}.
Package to use for differential expression calling. Defaults to \code{"edgeR"}
(faster for large datasets) but \code{"DESeq2"} is also supported.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells required to perform the differential expression
analysis.}

\item{minCellsPerGene}{\code{integer(1)}.
The minimum number of cells where a gene is expressed, to pass low
expression filtering.}

\item{minCountsPerCell}{\code{integer(1)}.
Minimum number of counts per cell for a gene to pass low expression
filtering. The number of cells is defined by \code{minCellsPerGene}.}

\item{bpparam}{\code{bpparamClass}.
BiocParallel parameter specifying the back-end to be used for computations.

We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
Varies depending on the \code{caller} argument:
\itemize{
\item \code{caller = "edgeR"}: \code{DEGLRT}.
\item \code{caller = "DESeq2"}: Unshrunken \code{DESeqResults}. Use \code{lfcShrink()} if
shrunken results are desired.
}
}
\description{
Differential expression
}
\details{
Perform pairwise differential expression across groups of cells by fitting
to a zero-inflated negative binomial (ZINB) model using the zinbwave package.
Currently supports edgeR and DESeq2 as DE callers.
}
\note{
We are currently recommending the ZINB-WaVE method over zingeR, since
it is faster, and has been show to be more sensitive for most single-cell
RNA-seq datasets.
}
\section{zinbwave}{


This function will run a lot faster if you pre-calculate the ZINB weights
using the \code{\link[=runZinbwave]{runZinbwave()}} function, which will stash the weights into the
\code{\link[SingleCellExperiment:weights]{SingleCellExperiment::weights()}} slot of the object. Running zinbwave across
the entire set of filtered cells also has greater sensitivity for weight
calculations.

We are currently using an epsilon setting of \code{1e12}, as recommended by the
ZINB-WaVE integration paper. For more information on the zinbwave package,
refer to these materials:
\itemize{
\item \href{https://doi.org/10.1186/s13059-018-1406-4}{zinbwave paper}.
\item \href{https://bit.ly/2wtDdpS}{zinbwave vignette}.
\item \href{https://github.com/mikelove/zinbwave-deseq2}{zinbwave-DESeq2 workflow}.
}
}

\section{edgeR}{


After estimation of the dispersions and posterior probabilities, the
\code{\link[zinbwave:glmWeightedF]{zinbwave::glmWeightedF()}} function is used for statistical inference. This
is an adaptation of \code{\link[edgeR:glmLRT]{edgeR::glmLRT()}}. It uses an F-test for which the
denominator degrees of freedom are by default adjusted according to the
downweighting of excess zeros (\code{ZI = TRUE}). Also, independent filtering can
be performed on the obtained p-values (\code{independentFiltering = TRUE}). We use
the independent filtering strategy that was originally implemented in DESeq2.
By default, the average fitted values are used as a filter criterion.
}

\section{DESeq2}{


We're providing preliminary support for DESeq2 as the differential expression
caller. It is currently considerably slower for large datasets than edgeR.

We're trying to follow the conventions used in DESeq2 for contrasts, defining
the name of the factor in the design formula, numerator, and denominator
level for the fold change calculations. See \code{\link[DESeq2:results]{DESeq2::results()}} for details.
}

\section{Seurat conventions}{


Note that Seurat currently uses the convention \code{cells.1} for the numerator
and \code{cells.2} for the denominator. See \code{\link[Seurat:DiffExpTest]{Seurat::DiffExpTest()}} for details.
}

\examples{
data(seurat_small)
object <- seurat_small

## Calculate ZINB weights (already stashed).
## seurat_small <- runZinbwave(object)

## Compare expression in cluster 3 relative to 2.
ident <- clusterID(object)
numerator <- names(ident)[ident == "3"]
summary(numerator)
denominator <- names(ident)[ident == "2"]
summary(denominator)

## edgeR.
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR"
)
class(x)
summary(x)

## DESeq2.
# This will warn about weights with the minimal example.
# x <- diffExp(
#     object = object,
#     numerator = numerator,
#     denominator = denominator,
#     caller = "DESeq2"
# )
# class(x)
# summary(x)
}
\seealso{
\code{Seurat::WhichCells()}.
}
