% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpPerCluster-methods.R
\docType{methods}
\name{diffExpPerCluster}
\alias{diffExpPerCluster}
\alias{diffExpPerCluster,SingleCellExperiment-method}
\alias{diffExpPerCluster,Seurat-method}
\title{Differential expression per cluster}
\usage{
diffExpPerCluster(object, ...)

\S4method{diffExpPerCluster}{SingleCellExperiment}(object, group,
  numerator, denominator, ...)

\S4method{diffExpPerCluster}{Seurat}(object, group, numerator, denominator,
  ...)
}
\arguments{
\item{object}{Object.}

\item{group}{\code{character(1)}.
Group of interest for differential expression per cluster. Must be a
\code{factor} column in \code{\link[SummarizedExperiment:colData]{colData()}}.}

\item{numerator}{\code{character}.
Cells to use in the numerator of the contrast (e.g. treatment).}

\item{denominator}{\code{character}.
Cells to use in the denominator of the contrast (e.g. control).}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Differential expression per cluster
}
\note{
Updated 2019-09-03.

Cluster identity (\code{ident}) must be defined in
\code{\link[SummarizedExperiment:colData]{colData()}} for this function to work.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
group <- factor(c("group1", "group2"))
colData(object)$group <- group
suppressMessages(
    x <- diffExpPerCluster(
        object = object,
        group = "group",
        numerator = "group2",
        denominator = "group1",
        caller = "edgeR"
    )
)
class(x)
lapply(x, class)
}
