% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffExpPerCluster-methods.R
\docType{methods}
\name{diffExpPerCluster}
\alias{diffExpPerCluster}
\alias{diffExpPerCluster,SingleCellExperiment-method}
\alias{diffExpPerCluster,seurat-method}
\title{Differential Expression per Cluster}
\usage{
diffExpPerCluster(object, ...)

\S4method{diffExpPerCluster}{SingleCellExperiment}(object, group,
  numerator, denominator, ...)

\S4method{diffExpPerCluster}{seurat}(object, group, numerator, denominator,
  ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Differential Expression per Cluster
}
\note{
Cluster identity (\code{ident}) must be defined in
\code{\link[SummarizedExperiment:colData]{SummarizedExperiment::colData()}} for this function to work.
}
\examples{
data(seurat_small)
object <- seurat_small
group <- factor(c("group1", "group2"))
colData(object)$group <- group
x <- suppressMessages(
    diffExpPerCluster(
        object = object,
        group = "group",
        numerator = "group2",
        denominator = "group1",
        caller = "edgeR"
    )
)
class(x)
lapply(x, class)
}
