% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSizeFactors-methods.R
\docType{methods}
\name{estimateSizeFactors}
\alias{estimateSizeFactors}
\alias{estimateSizeFactors,SingleCellExperiment-method}
\alias{estimateSizeFactors,cell_data_set-method}
\title{Estimate size factors}
\usage{
\S4method{estimateSizeFactors}{SingleCellExperiment}(object)

\S4method{estimateSizeFactors}{cell_data_set}(object)
}
\description{
Estimate size factors
}
\examples{
data(
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
sizeFactors(object)
object <- estimateSizeFactors(object)
sizeFactors(object)

## cell_data_set ====
object <- cell_data_set
object <- estimateSizeFactors(cell_data_set)
sizeFactors(object)
}
\seealso{
\itemize{
\item \code{DESeq2::estimateSizeFactors()}.
\item \code{monocle3::estimate_size_factors()}.
}
}
