% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers-methods.R
\docType{methods}
\name{findMarkers}
\alias{findMarkers}
\alias{findMarkers,SingleCellExperiment-method}
\alias{findMarkers,Seurat-method}
\title{Find cluster-specific marker genes}
\usage{
findMarkers(object, ...)

\S4method{findMarkers}{SingleCellExperiment}(object, ...)

\S4method{findMarkers}{Seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find cluster-specific marker genes
}
\note{
Cluster identity (\code{ident}) must be defined in
\code{\link[SummarizedExperiment:colData]{colData()}} for this function to work.
}
\examples{
data(seurat)
x <- findMarkers(seurat, caller = "edgeR")
class(x)
lapply(x, class)
}
