% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers-methods.R
\name{findMarkers}
\alias{findMarkers}
\alias{findMarkers,SingleCellExperiment-method}
\alias{findMarkers,Seurat-method}
\title{Find cluster-specific marker genes}
\usage{
findMarkers(object, ...)

\S4method{findMarkers}{SingleCellExperiment}(object, ...)

\S4method{findMarkers}{Seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find cluster-specific marker genes
}
\note{
Updated 2019-08-02.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
x <- findMarkers(object, caller = "edgeR")
class(x)
lapply(x, class)
}
