% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers-methods.R
\name{findMarkers}
\alias{findMarkers}
\alias{findMarkers,SingleCellExperiment-method}
\alias{findMarkers,Seurat-method}
\title{Find cluster-specific marker genes}
\usage{
findMarkers(object, ...)

\S4method{findMarkers}{SingleCellExperiment}(object, clusters = NULL, ...)

\S4method{findMarkers}{Seurat}(object, clusters = NULL, ...)
}
\arguments{
\item{object}{Object.}

\item{clusters}{\code{character} or \code{NULL}.
Cluster identifiers.
Must correspond to values in \code{\link[=clusters]{clusters()}}.
Note that Seurat uses zero-indexed IDs by default (e.g. 0, 1, 2, ...).
If left \code{NULL} (default), all clusters will be analyzed.
Looping across clusters manually here can avoid memory issues on laptops
and other machines with limited amounts of RAM.}

\item{...}{Passthrough arguments to \code{\link[=diffExp]{diffExp()}}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find cluster-specific marker genes
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
x <- findMarkers(object, caller = "edgeR")
class(x)
lapply(x, class)
}
