% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/findMarkers-methods.R
\docType{methods}
\name{findMarkers}
\alias{findMarkers}
\alias{findMarkers,SingleCellExperiment-method}
\alias{findMarkers,seurat-method}
\title{Find Cluster-Specific Marker Genes}
\usage{
findMarkers(object, ...)

\S4method{findMarkers}{SingleCellExperiment}(object, ...)

\S4method{findMarkers}{seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{diffExp()}.}
}
\value{
\code{list} containing:
\itemize{
\item \code{caller = "edgeR"}: \code{DGELRT}.
\item \code{caller = "DESeq2"}: \code{DESeqResults}.
}
}
\description{
Find Cluster-Specific Marker Genes
}
\note{
Cluster identity (\code{ident}) must be defined in \code{colData()} for this
function to work.
}
\examples{
data(seurat_small)
x <- findMarkers(seurat_small, caller = "edgeR")
class(x)
lapply(x, class)
}
