% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{importMarkers}
\alias{importMarkers}
\alias{importCellCycleMarkersFromGoogle}
\alias{importCellTypeMarkersFromGoogle}
\title{Import Markers}
\usage{
importCellCycleMarkersFromGoogle(gs, ws = 1L, gene2symbol = NULL,
  organism, genomeBuild, ensemblRelease)

importCellTypeMarkersFromGoogle(gs, ws = 1L, gene2symbol = NULL,
  organism, genomeBuild, ensemblRelease)
}
\arguments{
\item{gs}{\code{string}. Google Sheets \code{sheet_key} identifier, which
can be located with the \code{\link[googlesheets:gs_ls]{googlesheets::gs_ls()}} function.}

\item{ws}{positive integer or character string specifying index or title,
respectively, of the worksheet}

\item{gene2symbol}{\code{Gene2Symbol}. Gene-to-symbol mappings.}

\item{organism}{\code{string}. Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{ensemblRelease}{\code{scalar integer}. Ensembl release version (e.g. \code{90}).}

\item{file}{\code{string}. File path (CSV or Excel).}
}
\value{
\code{CellCycleMarkers} or \code{CellTypeMarkers}.
}
\description{
Currently designed for internal use by the pointillism package.
}
\section{Functions}{
\itemize{
\item \code{importCellCycleMarkersFromGoogle}: Cell-cycle markers.

\item \code{importCellTypeMarkersFromGoogle}: Cell-type markers.
}}

\section{Google Sheets}{


Google Sheets requires OAuth authentication.

See also:
\itemize{
\item \code{\link[googlesheets:gs_key]{googlesheets::gs_key()}}.
\item \code{\link[googlesheets:gs_read]{googlesheets::gs_read()}}.
}
}

\examples{
organism <- "Homo sapiens"
genomeBuild <- "GRCh38"
ensemblRelease <- 92L

## Google sheets mode currently must be interactive.
if (isTRUE(interactive())) {
    library(googlesheets)
    gs_ls()

    ## Gene-to-symbol mappings.
    g2s <- makeGene2SymbolFromEnsembl("Homo sapiens")

    ## Cell-cycle markers.
    x <- importCellCycleMarkersFromGoogle(
        gs = "1qA5ktYeimNGpZF1UPSQZATbpzEqgyxN6daoMOjv6YYw",
        ws = "Homo_sapiens",
        gene2symbol = g2s
    )

    ## Cell-type markers.
    x <- importCellTypeMarkersFromGoogle(
        gs = "1vGNU2CCxpaoTCLvzOxK1hf5gjULrf2-CpgCp9bOfGJ0",
        ws = "Homo_sapiens",
        gene2symbol = g2s
    )
}
}
