% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_markers_small}
\alias{known_markers_small}
\title{Known Markers Example}
\format{An object of class \code{KnownMarkers} with 5 rows and 10 columns.}
\usage{
known_markers_small
}
\description{
Known Markers Example
}
\examples{
data(known_markers_small)
known_markers_small
}
\keyword{datasets}
