% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KnownMarkers-class}
\alias{KnownMarkers-class}
\title{Known markers}
\value{
Results are grouped by \code{cellType} column and arranged by adjusted
\emph{P} value (\code{padj}).
}
\description{
Class containing known markers detected.
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
\concept{S4 classes}
