% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KnownMarkers-class}
\alias{KnownMarkers-class}
\title{Known Seurat Markers}
\value{
\code{KnownMarkers}.
}
\description{
Class containing essential elements for analysis of known, cell-type specific
marker genes.
}
\seealso{
\code{\link[=KnownMarkers]{KnownMarkers()}}.

Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{Markers-class}},
  \code{\link{MarkersPerCluster-class}}
}
\concept{S4 classes}
