% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/KnownMarkers-methods.R
\docType{class}
\name{KnownMarkers-class}
\alias{KnownMarkers-class}
\alias{KnownMarkers}
\alias{KnownMarkers,SeuratMarkersPerCluster,CellTypeMarkers-method}
\title{Known markers}
\usage{
KnownMarkers(markers, known, ...)

\S4method{KnownMarkers}{SeuratMarkersPerCluster,CellTypeMarkers}(markers,
  known, promiscuousThreshold = 0L)
}
\arguments{
\item{markers}{\code{SeuratMarkers} or \code{SeuratMarkersPerCluster}.}

\item{known}{\code{CellTypeMarkers}. Grouped by \code{cellType} column. Known markers
\code{data.frame} imported by \code{readCellTypeMarkers} or pulled from internal
cell cycle markers data.}

\item{...}{Additional arguments.}

\item{promiscuousThreshold}{\code{integer(1)}. Minimum number of clusters
required to consider a gene marker promiscuous. Set to \code{0} to disable
promiscuous marker filtering.}
}
\value{
\code{KnownMarkers}.
Results are grouped by \code{cellType} column and arranged by adjusted \emph{P} value
(\code{padj}).

\code{KnownMarkers}.
}
\description{
Class containing known markers detected.
}
\note{
Updated 2019-07-31.

Both the \code{markers} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
data(cellTypeMarkersList, seuratAllMarkers)

## SeuratMarkersPerCluster ====
markers <- seuratAllMarkers
known <- cellTypeMarkersList[["homoSapiens"]]

x <- KnownMarkers(
    markers = markers,
    known = known
)
summary(x)
}
