% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/knownMarkers-methods.R
\docType{methods}
\name{knownMarkers}
\alias{knownMarkers}
\alias{knownMarkers,Markers,CellTypeMarkers-method}
\title{Known Markers Detected}
\usage{
knownMarkers(all, known, ...)

\S4method{knownMarkers}{Markers,CellTypeMarkers}(all, known,
  promiscuousThreshold = 0L)
}
\arguments{
\item{all}{\code{Markers}.}

\item{known}{\code{CellTypeMarkers}. Grouped by \code{cellType} column. Known markers
\code{data.frame} imported by \code{\link[=readCellTypeMarkers]{readCellTypeMarkers()}} or pulled from internal
cell cycle markers data.}

\item{...}{Additional arguments.}

\item{promiscuousThreshold}{\code{scalar integer}. Minimum number of clusters
required to consider a gene marker promiscuous. Set to \code{0} to disable
promiscuous marker filtering.}
}
\value{
\code{KnownMarkers}.
}
\description{
Known Markers Detected
}
\note{
Both the \code{all} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
