% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/KnownMarkers-methods.R
\docType{methods}
\name{KnownMarkers}
\alias{KnownMarkers}
\alias{KnownMarkers,SeuratMarkersPerCluster,CellTypeMarkers-method}
\title{Known Markers}
\usage{
KnownMarkers(markers, known, ...)

\S4method{KnownMarkers}{SeuratMarkersPerCluster,CellTypeMarkers}(markers,
  known, promiscuousThreshold = 0L)
}
\arguments{
\item{markers}{\code{SeuratMarkers} or \code{SeuratMarkersPerCluster}.}

\item{known}{\code{CellTypeMarkers}. Grouped by \code{cellType} column. Known markers
\code{data.frame} imported by \code{\link[=readCellTypeMarkers]{readCellTypeMarkers()}} or pulled from internal
cell cycle markers data.}

\item{...}{Additional arguments.}

\item{promiscuousThreshold}{\code{scalar integer}. Minimum number of clusters
required to consider a gene marker promiscuous. Set to \code{0} to disable
promiscuous marker filtering.}
}
\value{
\code{KnownMarkers}.
}
\description{
Class containing known markers detected.
}
\note{
Both the \code{markers} and \code{known} objects must contain Ensembl gene
identifiers in the \code{geneID} column. We must avoid any matching operations
based on the gene names, since these change often and can mismatch
easily.
}
\examples{
data(all_markers_small, cell_type_markers)
x <- KnownMarkers(
    markers = all_markers_small,
    known = cell_type_markers$homoSapiens
)
summary(x)
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{SeuratMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
