% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-logcounts-methods.R
\docType{methods}
\name{logcounts}
\alias{logcounts}
\alias{logcounts,Seurat-method}
\alias{logcounts,cell_data_set-method}
\title{Log normalized counts}
\usage{
logcounts(object, ...)

\S4method{logcounts}{Seurat}(object, assay = NULL, verbose = FALSE)

\S4method{logcounts}{cell_data_set}(object, verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
\code{sparseMatrix}.
}
\description{
Log normalized counts
}
\note{
Updated 2019-08-05.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ===
object <- SingleCellExperiment
object <- normalize(object)
logcounts <- logcounts(object)
class(logcounts)

## Seurat ====
object <- Seurat
logcounts <- logcounts(object)
class(logcounts)

## cell_data_set ====
object <- cell_data_set
object <- normalize(object)
logcounts <- logcounts(object)
class(logcounts)
}
\seealso{
\itemize{
\item \code{\link[=normalize]{normalize()}}.
\item \code{scater::normalizeSCE()}.
\item \code{Seurat::NormalizeData()}.
\item \code{monocle3::prepare_cds()}.
}
}
\keyword{internal}
