% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-normcounts-methods.R
\docType{methods}
\name{normcounts}
\alias{normcounts}
\alias{normcounts,Seurat-method}
\alias{normcounts,cell_data_set-method}
\title{Size factor adjusted normalized counts}
\usage{
normcounts(object, ...)

\S4method{normcounts}{Seurat}(object, assay = NULL, verbose = FALSE)

\S4method{normcounts}{cell_data_set}(object, verbose = FALSE)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{verbose}{\code{logical(1)}.
Run the function with verbose output.}
}
\value{
\code{sparseMatrix}.
}
\description{
Size factor adjusted normalized counts
}
\note{
Updated 2019-08-05.
}
\examples{
data(
    Seurat,
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ===
object <- SingleCellExperiment
object <- normalize(object)
normcounts <- normcounts(object)
class(normcounts)

## Seurat ====
object <- Seurat
normcounts <- normcounts(object)
class(normcounts)

## cell_data_set ====
object <- cell_data_set
object <- normalize(object)
normcounts <- normcounts(object)
class(normcounts)
}
\seealso{
\itemize{
\item \code{\link[=normalize]{normalize()}}.
\item \code{scater::normalizeSCE()}.
\item \code{scater::normalizeCounts()}.
\item \code{Seurat::NormalizeData()}.
\item \code{monocle3::prepare_cds()}.
\item \code{monocle3::normalized_counts()}.
}
}
\keyword{internal}
