% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCountsPerCluster-methods.R
\name{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster,SingleCellExperiment-method}
\alias{plotCellCountsPerCluster,Seurat-method}
\alias{plotCellCountsPerCluster,cell_data_set-method}
\title{Plot cell counts per cluster}
\usage{
plotCellCountsPerCluster(object, ...)

\S4method{plotCellCountsPerCluster}{SingleCellExperiment}(object, interestingGroups = NULL)

\S4method{plotCellCountsPerCluster}{Seurat}(object, interestingGroups = NULL)

\S4method{plotCellCountsPerCluster}{cell_data_set}(object, interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output. Invisibly return \code{ggplot}.
}
\description{
Plot cell counts per cluster
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
plotCellCountsPerCluster(object)
}
