% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCountsPerCluster-methods.R
\docType{methods}
\name{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster}
\alias{plotCellCountsPerCluster,SingleCellExperiment-method}
\alias{plotCellCountsPerCluster,Seurat-method}
\title{Plot cell counts per cluster}
\usage{
\S4method{plotCellCountsPerCluster}{SingleCellExperiment}(object,
  interestingGroups = NULL)

\S4method{plotCellCountsPerCluster}{Seurat}(object,
  interestingGroups = NULL)
}
\arguments{
\item{object}{Object.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}
}
\value{
Show graphical output. Invisibly return \code{ggplot}.
}
\description{
Plot cell counts per cluster
}
\examples{
data(seurat)
plotCellCountsPerCluster(seurat)
}
