% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellTypesPerCluster-methods.R
\docType{methods}
\name{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster}
\alias{plotCellTypesPerCluster,SingleCellExperiment-method}
\alias{plotCellTypesPerCluster,seurat-method}
\title{Plot cell types per cluster}
\usage{
\S4method{plotCellTypesPerCluster}{SingleCellExperiment}(object, markers,
  min = 1L, max = Inf, reducedDim = "TSNE", expression = c("mean",
  "sum"), headerLevel = 2L, progress = FALSE, ...)

\S4method{plotCellTypesPerCluster}{seurat}(object, markers, min = 1L,
  max = Inf, reducedDim = "TSNE", expression = c("mean", "sum"),
  headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{KnownMarkers}.}

\item{min}{\code{numeric(1)}.
Recommended minimum value cutoff.}

\item{max}{\code{numeric(1)}.
Recommended maximum value cutoff.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\link[SingleCellExperiment:reducedDims]{reducedDims}. Includes TNSE, UMAP, PCA,
for example.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \link[base:match.arg]{match.arg()} internally and defaults to the first
argument in the \code{character} vector.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output. Invisibly return \code{list}.
}
\description{
Plot cell types per cluster
}
\details{
Plot the geometric mean of the significant marker genes for every known cell
type (per unbiased cluster). Cell types with too few (\code{min} cutoff) or too
many (\code{max} cutoff) marker genes will be skipped.
}
\examples{
data(seurat_small, known_markers_small)
plotCellTypesPerCluster(
    object = seurat_small,
    markers = known_markers_small
)
}
