% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts-methods.R, R/plotDot-methods.R,
%   R/plotViolin-methods.R
\docType{methods}
\name{plotCounts}
\alias{plotCounts}
\alias{plotDot}
\alias{plotViolin}
\alias{plotCounts,SingleCellExperiment-method}
\alias{plotCounts,Seurat-method}
\alias{plotCounts,cell_data_set-method}
\alias{plotDot,SingleCellExperiment-method}
\alias{plotDot,Seurat-method}
\alias{plotDot,cell_data_set-method}
\alias{plotViolin,SingleCellExperiment-method}
\alias{plotViolin,Seurat-method}
\alias{plotViolin,cell_data_set-method}
\title{Plot counts}
\usage{
plotCounts(object, ...)

\S4method{plotCounts}{SingleCellExperiment}(object, genes,
  value = c("logcounts", "normcounts"), geom = c("violin", "dot"),
  perSample = TRUE, legend = getOption("acid.legend", default = TRUE),
  title = NULL)

\S4method{plotCounts}{Seurat}(object, genes, value = c("logcounts",
  "normcounts"), geom = c("violin", "dot"), perSample = TRUE,
  legend = getOption("acid.legend", default = TRUE), title = NULL)

\S4method{plotCounts}{cell_data_set}(object, genes,
  value = c("logcounts", "normcounts"), geom = c("violin", "dot"),
  perSample = TRUE, legend = getOption("acid.legend", default = TRUE),
  title = NULL)

plotDot(object, ...)

\S4method{plotDot}{SingleCellExperiment}(object, genes, perSample = TRUE,
  colMin = -2.5, colMax = 2.5, dotMin = 0L, dotScale = 6L,
  color = getOption("acid.continuous.color", default =
  ggplot2::scale_colour_gradient2(low = "orange", mid = "gray75", high =
  "purple", midpoint = 0L)), legend = getOption("acid.legend", default =
  TRUE), title = NULL)

\S4method{plotDot}{Seurat}(object, genes, perSample = TRUE,
  colMin = -2.5, colMax = 2.5, dotMin = 0L, dotScale = 6L,
  color = getOption("acid.continuous.color", default =
  ggplot2::scale_colour_gradient2(low = "orange", mid = "gray75", high =
  "purple", midpoint = 0L)), legend = getOption("acid.legend", default =
  TRUE), title = NULL)

\S4method{plotDot}{cell_data_set}(object, genes, perSample = TRUE,
  colMin = -2.5, colMax = 2.5, dotMin = 0L, dotScale = 6L,
  color = getOption("acid.continuous.color", default =
  ggplot2::scale_colour_gradient2(low = "orange", mid = "gray75", high =
  "purple", midpoint = 0L)), legend = getOption("acid.legend", default =
  TRUE), title = NULL)

plotViolin(object, ...)

\S4method{plotViolin}{SingleCellExperiment}(object, genes,
  value = c("logcounts", "normcounts"), perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption("acid.discrete.color", default = NULL),
  legend = getOption("acid.legend", default = TRUE), title = NULL)

\S4method{plotViolin}{Seurat}(object, genes, value = c("logcounts",
  "normcounts"), perSample = TRUE, scale = c("count", "width", "area"),
  color = getOption("acid.discrete.color", default = NULL),
  legend = getOption("acid.legend", default = TRUE), title = NULL)

\S4method{plotViolin}{cell_data_set}(object, genes,
  value = c("logcounts", "normcounts"), perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption("acid.discrete.color", default = NULL),
  legend = getOption("acid.legend", default = TRUE), title = NULL)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers. It is considered better practice to input the stable gene
identifiers from Ensembl (e.g. "ENSG00000000003") and not the (HGNC) gene
symbols (e.g. "TSPN6"), if possible.}

\item{value}{Value to assign.}

\item{geom}{\code{character(1)}.
Plot type. Uses \code{\link[base:match.arg]{match.arg()}} to pick the type.
Currently supports \code{"dot"} and \code{"violin"}.}

\item{perSample}{\code{logical(1)}.
Visualize the distributions per sample.}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{title}{\code{character(1)}.
Plot title.}

\item{colMin}{\code{numeric(1)}.
Minimum scaled average expression threshold. Everything smaller will be
set to this.}

\item{colMax}{\code{numeric(1)}.
Maximum scaled average expression threshold. Everything larger will be set
to this.}

\item{dotMin}{\code{numeric(1)}.
The fraction of cells at which to draw the smallest dot. All cell groups
with less than this expressing the given gene will have no dot drawn.}

\item{dotScale}{\code{numeric(1)}.
Scale the size of the points, similar to \code{cex}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{scale}{\code{character(1)}.
Whether the values should be centered and scaled in either the row or
column direction ("\code{row}", "\code{column}"), or remain unscaled ("\code{none}").}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Visualize genes on a dot or violin plot.
}
\note{
Dot geom currently only supports logcounts.

Updated 2019-08-06.
}
\examples{
data(
    Seurat,
    cell_data_set,
    package = "acidtest"
)

## Seurat ====
object <- Seurat

## Plotting with either gene IDs or gene names (symbols) works.
genes <- head(rownames(object), n = 4L)
print(genes)

## Per sample mode enabled.
plotCounts(object, genes = genes, perSample = TRUE)

## Per sample mode disabled.
plotCounts(object, genes = genes, perSample = FALSE)

## cell_data_set ====
object <- cell_data_set
genes <- head(rownames(object), n = 4L)
print(genes)
plotCounts(object, genes = genes)
}
\seealso{
\itemize{
\item \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.
\item \code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}.
\item \code{\link[Seurat:RidgePlot]{Seurat::RidgePlot()}}.
\item \code{\link[monocle3:plot_genes_violin]{monocle3::plot_genes_violin()}}.
}
}
