% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeature-methods.R
\name{plotFeature}
\alias{plotFeature}
\title{Plot Feature}
\usage{
plotFeature(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
object <- sce_small
features <- c("nUMI", "nGene", "mitoRatio")
plotFeature(object, features = features)
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.

Other Clustering Functions: \code{\link{cellTypesPerCluster}},
  \code{\link{knownMarkersDetected}},
  \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{sanitizeSeuratMarkers}},
  \code{\link{topMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Clustering Functions}
