% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature-methods.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,SingleCellExperiment-method}
\alias{plotFeature,Seurat-method}
\title{Plot feature}
\usage{
plotFeature(object, ...)

\S4method{plotFeature}{SingleCellExperiment}(object, features,
  reduction = "UMAP", color = getOption("acid.continuous.color",
  default = ggplot2::scale_colour_gradient(low = "gray75", high =
  "purple")), pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  expression = c("mean", "sum"))

\S4method{plotFeature}{Seurat}(object, features, reduction = "UMAP",
  color = getOption("acid.continuous.color", default =
  ggplot2::scale_colour_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  expression = c("mean", "sum"))
}
\arguments{
\item{object}{Object.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{reduction}{\code{character(1)}.
Name of dimension reduction matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes UMAP, TSNE, PCA, typically, for example.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level. Useful when there many points in the dataset
(e.g. single-cell data), and some points can be masked.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the
\code{\link[acidplots:theme_midnight]{theme_midnight()}} ggplot2 theme.}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} (1 feature) or \code{list} (multiple features).
}
\description{
Plot feature
}
\note{
Updated 2019-07-31.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
plotFeature(object, features = c("nCount_RNA", "nFeature_RNA"))
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
