% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotFeature-methods.R
\docType{methods}
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,SingleCellExperiment-method}
\alias{plotFeature,seurat-method}
\title{Plot Feature}
\usage{
plotFeature(object, ...)

\S4method{plotFeature}{SingleCellExperiment}(object, features,
  reducedDim = "TSNE",
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE),
  expression = c("mean", "sum"))

\S4method{plotFeature}{seurat}(object, features, reducedDim = "TSNE",
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE),
  expression = c("mean", "sum"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.discrete.color = ggplot2::scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{basejump::theme_midnight()} ggplot2 theme.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{\link[base:match.arg]{base::match.arg()}}
and defaults to the first argument in the \code{character} vector.}
}
\value{
\code{ggplot} or \code{list}.
}
\description{
Plot Feature
}
\examples{
data(seurat_small)
plotFeature(seurat_small, features = c("nUMI", "nGene"))
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
