% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature-methods.R
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,SingleCellExperiment-method}
\alias{plotFeature,Seurat-method}
\title{Plot feature}
\usage{
plotFeature(object, ...)

\S4method{plotFeature}{SingleCellExperiment}(
  object,
  features,
  reduction = "UMAP",
  color = getOption("acid.continuous.color", default = ggplot2::scale_color_gradient(low
    = "gray75", high = "purple")),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  expression = c("mean", "sum")
)

\S4method{plotFeature}{Seurat}(
  object,
  features,
  reduction = "UMAP",
  color = getOption("acid.continuous.color", default = ggplot2::scale_color_gradient(low
    = "gray75", high = "purple")),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  expression = c("mean", "sum")
)
}
\arguments{
\item{object}{Object.}

\item{features}{\code{character}. Features to plot (e.g. gene expression, PC
scores, number of genes detected).}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{acid_theme_light()} ggplot2
theme.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot} (1 feature) or \code{list} (multiple features).
}
\description{
Plot feature
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "acidtest")

## Seurat ====
object <- Seurat
plotFeature(
    object = object,
    features = c("nCount_RNA", "nFeature_RNA", "PC_1", "PC_2"),
    reduction = "UMAP"
)
}
\seealso{
\code{\link[Seurat:FeaturePlot]{Seurat::FeaturePlot()}}.
}
