% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotGene-methods.R
\docType{methods}
\name{plotDot}
\alias{plotDot}
\alias{plotViolin}
\alias{plotGene}
\alias{plotGene,SingleCellExperiment-method}
\alias{plotGene,seurat-method}
\alias{plotDot,SingleCellExperiment-method}
\alias{plotDot,seurat-method}
\alias{plotViolin,SingleCellExperiment-method}
\alias{plotViolin,seurat-method}
\title{Plot Gene}
\usage{
plotDot(object, ...)

plotViolin(object, ...)

plotGene(object, ...)

\S4method{plotGene}{SingleCellExperiment}(object, genes, geom = c("dot",
  "violin"), perSample = TRUE,
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotGene}{seurat}(object, genes, geom = c("dot", "violin"),
  perSample = TRUE, color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotDot}{SingleCellExperiment}(object, genes, perSample = TRUE,
  colMin = -2.5, colMax = 2.5, dotMin = 0L, dotScale = 6L,
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "orange", high = "purple")),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotDot}{seurat}(object, genes, perSample = TRUE,
  colMin = -2.5, colMax = 2.5, dotMin = 0L, dotScale = 6L,
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_color_gradient(low = "orange", high = "purple")),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotViolin}{SingleCellExperiment}(object, genes,
  perSample = TRUE, scale = c("count", "width", "area"),
  color = getOption("pointillism.discrete.color", NULL),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotViolin}{seurat}(object, genes, perSample = TRUE,
  scale = c("count", "width", "area"),
  color = getOption("pointillism.discrete.color", NULL),
  legend = getOption("pointillism.legend", TRUE), title = NULL)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{geom}{\code{string}. Plot type. Uses \code{\link[=match.arg]{match.arg()}} to pick the type.
Currently supports "\code{dot}" and "\code{violin}".}

\item{perSample}{\code{boolean}. Visualize the distributions per sample?}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{colMin}{\code{scalar numeric}. Minimum scaled average expression threshold.
Everything smaller will be set to this.}

\item{colMax}{\code{scalar numeric}. Maximum scaled average expression threshold.
Everything larger will be set to this.}

\item{dotMin}{\code{scalar numeric}. The fraction of cells at which to draw the
smallest dot. All cell groups with less than this expressing the given gene
will have no dot drawn.}

\item{dotScale}{\code{scalar numeric}. Scale the size of the points, similar to
\code{cex}.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}
}
\value{
\code{ggplot}.
}
\description{
Visualize genes on a dot or violin plot.
}
\examples{
data(seurat_small)
object <- seurat_small

## Plotting with either gene IDs or gene names (symbols) works.
geneIDs <- head(rownames(object), n = 4L)
print(geneIDs)
geneNames <- head(as.character(rowRanges(object)$geneName), n = 4L)
print(geneNames)

## Per sample mode enabled.
plotDot(object, genes = geneNames, perSample = TRUE)
plotViolin(object, genes = geneNames, perSample = TRUE)

## Per sample mode disabled.
plotDot(object, genes = geneIDs, perSample = FALSE)
plotViolin(object, genes = geneIDs, perSample = FALSE)
}
\seealso{
\itemize{
\item \code{\link[Seurat:DotPlot]{Seurat::DotPlot()}}.
\item \code{\link[Seurat:VlnPlot]{Seurat::VlnPlot()}}.
\item \code{\link[Seurat:RidgePlot]{Seurat::RidgePlot()}}.
}

Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotFeature}},
  \code{\link{plotKnownMarkers}}, \code{\link{plotMarker}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\concept{Plot Functions}
