% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnownMarkers-methods.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownMarkers-method}
\alias{plotKnownMarkers,seurat,KnownMarkers-method}
\title{Plot known markers}
\usage{
\S4method{plotKnownMarkers}{SingleCellExperiment,KnownMarkers}(object,
  markers, reducedDim = "TSNE", headerLevel = 2L, progress = FALSE,
  ...)

\S4method{plotKnownMarkers}{seurat,KnownMarkers}(object, markers,
  reducedDim = "TSNE", headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{grouped_df}.
Marker genes, grouped by \code{"cellType"}.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\link[SingleCellExperiment:reducedDims]{reducedDims}. Includes TNSE, UMAP, PCA,
for example.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot known markers
}
\examples{
data(seurat_small, known_markers_small)
plotKnownMarkers(
    object = seurat_small,
    markers = known_markers_small
)
}
