% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKnownMarkers-methods.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownMarkers-method}
\alias{plotKnownMarkers,Seurat,KnownMarkers-method}
\title{Plot known markers}
\usage{
plotKnownMarkers(object, markers, ...)

\S4method{plotKnownMarkers}{SingleCellExperiment,KnownMarkers}(object,
  markers, reduction = "UMAP", headerLevel = 2L, progress = FALSE,
  ...)

\S4method{plotKnownMarkers}{Seurat,KnownMarkers}(object, markers,
  reduction = "UMAP", headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot known markers
}
\note{
Updated 2019-08-02.
}
\examples{
data(Seurat, package = "acidtest")
data(seuratKnownMarkers)

## Seurat ====
object <- Seurat
markers <- seuratKnownMarkers
plotKnownMarkers(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
