% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotKnownMarkers-methods.R
\docType{methods}
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownMarkers-method}
\alias{plotKnownMarkers,seurat,KnownMarkers-method}
\title{Plot Known Markers}
\usage{
plotKnownMarkers(object, markers, known, ...)

\S4method{plotKnownMarkers}{SingleCellExperiment,KnownMarkers}(object,
  markers, reducedDim = "TSNE", headerLevel = 2L, progress = FALSE,
  ...)

\S4method{plotKnownMarkers}{seurat,KnownMarkers}(object, markers,
  reducedDim = "TSNE", headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{grouped_df}. Marker genes, grouped by "\code{cellType}".}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{progress}{\code{boolean}. Show progress, using progress bars.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Plot Known Markers
}
\examples{
data(seurat_small, known_markers_small)
plotKnownMarkers(
    object = seurat_small,
    markers = known_markers_small
)
}
