% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotKnownMarkers-methods.R
\name{plotKnownMarkers}
\alias{plotKnownMarkers}
\alias{plotKnownMarkers,SingleCellExperiment,KnownMarkers-method}
\alias{plotKnownMarkers,Seurat,KnownMarkers-method}
\title{Plot known markers}
\usage{
plotKnownMarkers(object, markers, ...)

\S4method{plotKnownMarkers}{SingleCellExperiment,KnownMarkers}(
  object,
  markers,
  reduction = "UMAP",
  headerLevel = 2L,
  ...,
  BPPARAM = BiocParallel::bpparam()
)

\S4method{plotKnownMarkers}{Seurat,KnownMarkers}(
  object,
  markers,
  reduction = "UMAP",
  headerLevel = 2L,
  ...,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot known markers
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "AcidTest")
data(seuratKnownMarkers)

## Seurat ====
object <- Seurat
markers <- seuratKnownMarkers
plotKnownMarkers(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
