% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMarker-methods.R
\docType{methods}
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarker,SingleCellExperiment-method}
\alias{plotMarker,seurat-method}
\title{Plot Cell-Type-Specific Gene Markers}
\usage{
plotMarker(object, ...)

\S4method{plotMarker}{SingleCellExperiment}(object, genes,
  reducedDim = "TSNE", expression = c("mean", "sum"),
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_colour_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)

\S4method{plotMarker}{seurat}(object, genes, reducedDim = "TSNE",
  expression = c("mean", "sum"),
  color = getOption("pointillism.continuous.color",
  ggplot2::scale_colour_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. It is considered better practice
to input the stable gene identifiers from Ensembl (e.g. "ENSG00000000003")
and not the (HGNC) gene symbols (e.g. "TSPN6"), if possible.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{SingleCellExperiment::reducedDims()}. Includes TNSE, UMAP, PCA, for
example.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{base::match.arg()}
and defaults to the first argument in the \code{character} vector.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{ggplot2::scale_color_manual()}. To set the discrete color
palette globally, use
\code{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Point size for dots in the plot.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many points in the dataset (e.g. single-cell data), and
some points can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Superimpose sample text labels on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{basejump::theme_midnight()} ggplot2 theme.}

\item{legend}{\code{boolean}. Show plot legend.}

\item{title}{\code{string}. Plot title.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\examples{
data(seurat_small)
object <- seurat_small
title <- "most abundant genes"
genes <- counts(object) \%>\%
    Matrix::rowSums(.) \%>\%
    sort(decreasing = TRUE) \%>\%
    head(n = 4L) \%>\%
    names()
str(genes)

## Default appearance.
plotMarker(object, genes = genes[[1L]])

## Dark mode with viridis palette.
plotMarker(
    object = object,
    genes = genes,
    expression = "mean",
    pointsAsNumbers = TRUE,
    color = ggplot2::scale_color_viridis_c(),
    dark = TRUE,
    label = FALSE,
    title = title
)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
