% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotMarker-methods.R
\docType{methods}
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarker,SingleCellExperiment-method}
\alias{plotMarker,seurat-method}
\title{Plot Cell-Type-Specific Gene Markers}
\usage{
plotMarker(object, ...)

\S4method{plotMarker}{SingleCellExperiment}(object, genes,
  reducedDim = c("TSNE", "UMAP", "PCA"), expression = c("mean", "sum"),
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.8),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)

\S4method{plotMarker}{seurat}(object, genes, reducedDim = c("TSNE",
  "UMAP", "PCA"), expression = c("mean", "sum"),
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.8),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = TRUE)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{\link[=match.arg]{match.arg()}} and
defaults to the first argument in the \code{character} vector.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{title}{\code{string} or \code{NULL}. Plot title.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\examples{
title <- "most abundant genes"
genes <- counts(sce_small) \%>\%
    Matrix::rowSums(.) \%>\%
    sort(decreasing = TRUE) \%>\%
    head(n = 4L) \%>\%
    names()
glimpse(genes)
plotMarker(sce_small, genes = genes[[1L]])
plotMarker(
    object = sce_small,
    genes = genes,
    expression = "mean",
    pointsAsNumbers = TRUE,
    dark = TRUE,
    label = FALSE,
    title = title
)
}
\seealso{
Other Plot Functions: \code{\link{plotCellTypesPerCluster}},
  \code{\link{plotDot}}, \code{\link{plotFeature}},
  \code{\link{plotKnownMarkersDetected}},
  \code{\link{plotPCElbow}}, \code{\link{plotReducedDim}},
  \code{\link{plotTopMarkers}}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
\concept{Plot Functions}
