% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarker-methods.R
\name{plotMarker}
\alias{plotMarker}
\alias{plotMarker,SingleCellExperiment-method}
\alias{plotMarker,Seurat-method}
\title{Plot cell-type-specific gene markers}
\usage{
plotMarker(object, ...)

\S4method{plotMarker}{SingleCellExperiment}(
  object,
  genes,
  reduction = "UMAP",
  expression = c("mean", "sum"),
  color = getOption("acid.continuous.color", default =
    ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  labels = list(title = "auto", subtitle = NULL)
)

\S4method{plotMarker}{Seurat}(
  object,
  genes,
  reduction = "UMAP",
  expression = c("mean", "sum"),
  color = getOption("acid.continuous.color", default =
    ggplot2::scale_color_gradient(low = "gray75", high = "purple")),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  labels = list(title = "auto", subtitle = NULL)
)
}
\arguments{
\item{object}{Object.}

\item{genes}{\code{character}.
Gene identifiers.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{expression}{\code{character(1)}.
Calculation to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults to the first
argument in the \code{character} vector.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{acid_theme_light()} ggplot2
theme.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
Show graphical output.
Invisibly return a \code{ggplot} \code{list}.
}
\description{
Visualize gene markers on a reduced dimension plot (e.g. t-SNE, UMAP).
}
\note{
Updated 2020-02-21.
}
\examples{
data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
genes <- counts(object) \%>\%
    Matrix::rowSums(.) \%>\%
    sort(decreasing = TRUE) \%>\%
    head(n = 4L) \%>\%
    names()
print(genes)
plotMarker(
    object = object,
    genes = genes,
    reduction = "UMAP"
)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
