% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCElbow-methods.R
\docType{methods}
\name{plotPCElbow}
\alias{plotPCElbow}
\alias{plotPCElbow,Seurat-method}
\alias{plotPCElbow,cell_data_set-method}
\title{Plot principal component elbow}
\usage{
plotPCElbow(object, ...)

\S4method{plotPCElbow}{Seurat}(object, reduction = "PCA", minSD = 1L,
  minPct = 0.01, maxCumPct = 0.9, trans = c("identity", "sqrt"))

\S4method{plotPCElbow}{cell_data_set}(object)
}
\arguments{
\item{object}{Object.}

\item{reduction}{\code{character(1)}.
Name of dimension reduction matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes UMAP, TSNE, PCA, typically, for example.}

\item{minSD}{\code{numeric(1)}.
Minimum standard deviation.}

\item{minPct}{\code{numeric(1)} (\code{0}-\code{1}).
Minimum percent standard deviation.}

\item{maxCumPct}{\code{numeric(1)} (\code{0}-\code{1}).
Maximum cumulative percen standard deviation.}

\item{trans}{\code{character(1)}.
Name of the axis scale transformation to apply.

For more information:\preformatted{help(topic = "scale_x_continuous", package = "ggplot2")
}}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item Show graphical output of elbow plots.
\item Invisibly return numeric sequence vector of PCs to include for
dimensionality reduction analysis.
}
}
\description{
Calculate the principal component (PC) cutoff using a heuristic approach.
}
\details{
Automatically return the smallest number of PCs that match the \code{minSD},
\code{minPct}, and \code{maxCumPct} cutoffs.
}
\note{
Updated 2019-08-02.
}
\examples{
data(
    Seurat,
    cell_data_set,
    package = "acidtest"
)

## Seurat ====
object <- Seurat
plotPCElbow(object)

## cell_data_set ====
object <- cell_data_set
plotPCElbow(object)
}
\seealso{
\itemize{
\item \code{Seurat::PCElbowPlot()}.
\item \code{monocle3::plot_pc_variance_explained()}.
}
}
