% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDim-methods.R
\name{plotReducedDim}
\alias{plotReducedDim}
\alias{plotPCA}
\alias{plotTSNE}
\alias{plotUMAP}
\alias{plotReducedDim,SingleCellExperiment-method}
\alias{plotReducedDim,Seurat-method}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA,Seurat-method}
\alias{plotTSNE,SingleCellExperiment-method}
\alias{plotTSNE,Seurat-method}
\alias{plotUMAP,SingleCellExperiment-method}
\alias{plotUMAP,Seurat-method}
\title{Plot reduced dimensions}
\usage{
plotReducedDim(object, ...)

plotPCA(object, ...)

plotTSNE(object, ...)

plotUMAP(object, ...)

\S4method{plotReducedDim}{SingleCellExperiment}(
  object,
  reduction = "UMAP",
  dims = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("acid.discrete.color", default =
    AcidPlots::scale_color_synesthesia_d()),
  pointSize = getOption("acid.pointSize", default = 0.75),
  pointAlpha = getOption("acid.pointAlpha", default = 0.85),
  pointsAsNumbers = getOption("acid.pointsAsNumbers", default = FALSE),
  label = getOption("acid.label", default = TRUE),
  labelSize = getOption("acid.labelSize", default = 6L),
  dark = getOption("acid.dark", default = FALSE),
  legend = getOption("acid.legend", default = TRUE),
  labels = list(title = NULL, subtitle = NULL)
)

\S4method{plotReducedDim}{Seurat}(object, ...)

\S4method{plotPCA}{SingleCellExperiment}(object, ...)

\S4method{plotPCA}{Seurat}(object, ...)

\S4method{plotTSNE}{SingleCellExperiment}(object, ...)

\S4method{plotTSNE}{Seurat}(object, ...)

\S4method{plotUMAP}{SingleCellExperiment}(object, ...)

\S4method{plotUMAP}{Seurat}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{dims}{\code{integer}.
Vector of length 2 that denotes the columns from the reduced dimension
matrix to use for \code{centerX} and \code{centerY} column calculations. Defaults the
first and second dimensions.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{ggplot2::scale_color_manual()}.

To set the discrete color palette globally, use:\preformatted{options(acid.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.
In the range of 1-3 is generally recommended.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{acid_theme_light()} ggplot2
theme.}

\item{legend}{\code{logical(1)}.
Include plot legend.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot cell-level dimensionality reduction.
}
\details{
Colors using \code{ident} column defined in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} by default.
}
\note{
Updated 2020-02-21.
}
\section{Reduction types}{

\itemize{
\item PCA: \strong{P}rincipal \strong{C}omponent \strong{A}nalysis.
\item t-SNE: \strong{t}-distributed \strong{S}tochastic \strong{N}eighbor \strong{E}mbedding.
\item UMAP: \strong{U}niform \strong{M}anifold \strong{A}pproximation and \strong{P}rojection.
}
}

\section{UMAP calculation}{


\href{https://github.com/lmcinnes/umap}{UMAP} calculation in R requires the \href{https://www.python.org}{Python} module \code{umap-learn}.
The UMAP module can be loaded in R using \href{https://rstudio.github.io/reticulate/}{reticulate}.
}

\examples{
data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
plotReducedDim(object, reduction = "UMAP")
}
\seealso{
\itemize{
\item \code{Seurat::DimPlot()}.
\item \code{monocle3::plot_cells()}.
\item \href{https://satijalab.org/seurat/mca.html}{Seurat Mouse Cell Atlas vignette}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
