% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDim-methods.R
\docType{methods}
\name{plotReducedDim}
\alias{plotReducedDim}
\alias{plotReducedDim,SingleCellExperiment-method}
\alias{plotReducedDim,seurat-method}
\alias{plotTSNE,SingleCellExperiment-method}
\alias{plotTSNE,seurat-method}
\alias{plotUMAP,SingleCellExperiment-method}
\alias{plotUMAP,seurat-method}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA,seurat-method}
\title{Plot reduced dimensions}
\usage{
\S4method{plotReducedDim}{SingleCellExperiment}(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotReducedDim}{seurat}(object, reducedDim, dimsUse = c(1L,
  2L), interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotTSNE}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotTSNE}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotUMAP}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotUMAP}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotPCA}{SingleCellExperiment}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)

\S4method{plotPCA}{seurat}(object, dimsUse = c(1L, 2L),
  interestingGroups = NULL,
  color = getOption("pointillism.discrete.color", NULL),
  pointSize = getOption("pointillism.pointSize", 0.75),
  pointAlpha = getOption("pointillism.pointAlpha", 0.85),
  pointsAsNumbers = getOption("pointillism.pointsAsNumbers", FALSE),
  label = getOption("pointillism.label", TRUE),
  labelSize = getOption("pointillism.labelSize", 6L),
  dark = getOption("pointillism.dark", FALSE),
  legend = getOption("pointillism.legend", TRUE), title = NULL)
}
\arguments{
\item{object}{Object.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\link[SingleCellExperiment:reducedDims]{reducedDims}. Includes TNSE, UMAP, PCA,
for example.}

\item{dimsUse}{\code{integer}.
Vector of length 2 that denotes the columns from the reduced dimension
matrix to use for \code{centerX} and \code{centerY} column calculations. Defaults the
first and second dimensions.}

\item{interestingGroups}{\code{character}.
Groups of interest that define the samples. If left unset, defaults to
\code{sampleName}.}

\item{color}{\code{ScaleDiscrete}.
Desired ggplot2 color scale. Must supply discrete values. When set \code{NULL},
the default ggplot2 color palette will be used. If manual color definitions
are desired, we recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.

To set the discrete color palette globally, use:\preformatted{options(basejump.color.discrete = ggplot2::scale_color_viridis_d())
}}

\item{pointSize}{\code{numeric(1)}.
Point size for dots in the plot.}

\item{pointAlpha}{\code{numeric(1)} (\code{0}-\code{1}).
Alpha transparency level. Useful when there many points in the dataset
(e.g. single-cell data), and some points can be masked.}

\item{pointsAsNumbers}{\code{logical(1)}.
Plot the points as numbers (\code{TRUE}) or dots (\code{FALSE}).}

\item{label}{\code{logical(1)}.
Superimpose sample text labels on the plot.}

\item{labelSize}{\code{integer(1)}.
Size of the text label.}

\item{dark}{\code{logical(1)}.
Plot against a dark background using the \code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}
ggplot2 theme.}

\item{legend}{\code{logical(1)}.
Show plot legend.}

\item{title}{\code{character(1)}.
Plot title.}
}
\value{
\code{ggplot}.
}
\description{
Plot reduced dimensions
}
\details{
Colors using \code{ident} column defined in
\code{\link[SummarizedExperiment:colData]{colData()}} by default.
}
\section{Reduction types}{

\itemize{
\item PCA: \strong{P}rincipal \strong{C}omponent \strong{A}nalysis.
\item t-SNE: \strong{t}-distributed \strong{S}tochastic \strong{N}eighbor \strong{E}mbedding.
\item UMAP: \strong{U}niform \strong{M}anifold \strong{A}pproximation and \strong{P}rojection.
}
}

\section{UMAP calculation}{


\href{https://github.com/lmcinnes/umap}{UMAP} calculation in R requires the \href{https://www.python.org}{Python} module \code{umap-learn}.

We recommend installing this with \href{https://conda.io}{conda}:\preformatted{conda install -c conda-forge umap-learn
}

The UMAP module can be loaded in R using \href{https://rstudio.github.io/reticulate/}{reticulate}. Reticulate works
reliably when setting \code{RETICULATE_PYTHON} to point to your conda python
binary. Export this variable in  \code{~/.Renviron}.

See \code{base::Sys.getenv} for more information on the R system environment.
}

\examples{
data(seurat_small)
object <- seurat_small

## t-SNE
plotTSNE(object)
plotTSNE(object, pointsAsNumbers = TRUE, dark = TRUE, label = FALSE)

## UMAP
plotUMAP(object)

## PCA
plotPCA(object)
}
\seealso{
\itemize{
\item \code{Seurat::DimPlot}.
\item \href{https://satijalab.org/seurat/mca.html}{Seurat Mouse Cell Atlas vignette}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
