% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedBarPlot-methods.R
\name{plotStackedBarPlot}
\alias{plotStackedBarPlot}
\alias{plotStackedBarPlot,SingleCellExperiment-method}
\title{Plot stacked bar plot}
\usage{
\S4method{plotStackedBarPlot}{SingleCellExperiment}(
  object,
  absolute = FALSE,
  interestingGroups = NULL,
  labels = NULL
)
}
\arguments{
\item{object}{Object.}

\item{absolute}{\code{logical(1)}.
Return absolute (\code{TRUE}) or relative/proportional (\code{FALSE}) cell count.}

\item{interestingGroups}{\code{character}.
Groups of interest to use for visualization.
Corresponds to factors describing the columns of the object.}

\item{labels}{\code{list}.
ggplot2 labels.
See \code{\link[ggplot2:labs]{ggplot2::labs()}} for details.}

\item{...}{Additional arguments.}
}
\value{
\code{ggplot}.
}
\description{
Plot stacked bar plot
}
\note{
Updated 2021-03-04.
}
\examples{
data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
plotStackedBarPlot(object)
}
