% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTopMarkers-methods.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,SingleCellExperiment-method}
\alias{plotTopMarkers,seurat-method}
\title{Plot Top Markers}
\usage{
plotTopMarkers(object, ...)

\S4method{plotTopMarkers}{SingleCellExperiment}(object, markers, n = 1L,
  direction = c("up", "down", "both"), reducedDim = "TSNE",
  headerLevel = 2L, progress = FALSE, ...)

\S4method{plotTopMarkers}{seurat}(object, markers, n = 1L,
  direction = c("up", "down", "both"), reducedDim = "TSNE",
  headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}

\item{markers}{\code{grouped_df}. Marker genes, grouped by "\code{cluster}".}

\item{n}{\code{scalar integer}. Number to include.}

\item{direction}{\code{string}. Plot "\code{both}", "\code{up}", or "\code{down}" directions.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). Markdown header level.}

\item{progress}{\code{boolean}. Show progress, using progress bars.}
}
\value{
Show graphical output. Invisibly return \code{ggplot} \code{list}.
}
\description{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\examples{
data(seurat_small, all_markers_small)
plotTopMarkers(object = seurat_small, markers = all_markers_small)
}
