% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMarkers-methods.R
\docType{methods}
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,SingleCellExperiment-method}
\alias{plotTopMarkers,Seurat-method}
\title{Plot top markers}
\usage{
\S4method{plotTopMarkers}{SingleCellExperiment}(object, markers, n = 1L,
  direction = c("up", "down", "both"), reducedDim = "TSNE",
  headerLevel = 2L, progress = FALSE, ...)

\S4method{plotTopMarkers}{Seurat}(object, markers, n = 1L,
  direction = c("up", "down", "both"), reducedDim = "TSNE",
  headerLevel = 2L, progress = FALSE, ...)
}
\arguments{
\item{object}{Object.}

\item{markers}{\code{grouped_df}.
Marker genes, grouped by "\code{cluster}".}

\item{n}{\code{integer(1)}.
Number to include.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{reducedDim}{\code{character(1)}.
Name of reduced dimension matrix slotted in
\code{\link[SingleCellExperiment:reducedDims]{reducedDims()}}.
Includes TNSE, UMAP, PCA, for example.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{progress}{\code{logical(1)}.
Show progress, using progress bars.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot top markers
}
\details{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\examples{
data(seurat_small, all_markers_small)
plotTopMarkers(object = seurat_small, markers = all_markers_small)
}
