% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotTopMarkers-methods.R
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,Seurat,SeuratMarkersPerCluster-method}
\title{Plot top markers}
\usage{
plotTopMarkers(object, markers, ...)

\S4method{plotTopMarkers}{Seurat,SeuratMarkersPerCluster}(
  object,
  markers,
  direction = c("both", "up", "down"),
  reduction = "UMAP",
  n = 1L,
  headerLevel = 2L,
  ...,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing cell marker expression data.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Passthrough arguments to \code{plotMarker()}.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot top markers
}
\details{
The number of markers to plot is determined by the output of the
\code{topMarkers()} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\note{
Updated 2022-06-09.
}
\examples{
data(Seurat, SeuratMarkersPerCluster, package = "AcidTest")

## Seurat, SeuratMarkersPerCluster ====
object <- Seurat
markers <- SeuratMarkersPerCluster
plotTopMarkers(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
