% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopMarkers-methods.R
\name{plotTopMarkers}
\alias{plotTopMarkers}
\alias{plotTopMarkers,Seurat,SeuratMarkersPerCluster-method}
\title{Plot top markers}
\usage{
plotTopMarkers(object, markers, ...)

\S4method{plotTopMarkers}{Seurat,SeuratMarkersPerCluster}(
  object,
  markers,
  n = 1L,
  direction = c("up", "down", "both"),
  reduction = "UMAP",
  headerLevel = 2L,
  ...,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{object}{Object.}

\item{markers}{Object containing cell marker expression data.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{reduction}{\code{vector(1)}.
Dimension reduction name or index position.}

\item{headerLevel}{\code{integer(1)} (\code{1}-\code{7}).
Markdown header level.}

\item{...}{Passthrough arguments to \code{\link[=plotMarker]{plotMarker()}}.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
Show graphical output. Invisibly return a \code{ggplot} \code{list}.
}
\description{
Plot top markers
}
\details{
The number of markers to plot is determined by the output of the
\code{\link[=topMarkers]{topMarkers()}} function. If you want to reduce the number of genes to plot,
simply reassign first using that function. If necessary, we can add support
for the number of genes to plot here in a future update.
}
\note{
Updated 2020-01-30.
}
\examples{
data(Seurat, package = "acidtest")
data(seurat_all_markers)

## Seurat, SeuratMarkersPerCluster ====
object <- Seurat
markers <- seurat_all_markers
plotTopMarkers(
    object = object,
    markers = markers,
    reduction = "UMAP"
)
}
