% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pointillism-package}
\alias{pointillism}
\alias{pointillism-package}
\title{pointillism}
\description{
R package for for single-cell RNA-seq clustering analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://steinbaugh.com/pointillism}
  \item \url{https://github.com/steinbaugh/pointillism}
  \item Report bugs at \url{https://github.com/steinbaugh/pointillism/issues}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Rory Kirchner \email{roryk@alum.mit.edu} (0000-0003-4814-5885) [contributor]
}

}
\keyword{internal}
