% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene-methods.R, R/show-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{basejump}{\code{\link[basejump]{plotGene}}}

  \item{methods}{\code{\link[methods]{show}}}
}}

