% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeurat.R
\name{runSeurat}
\alias{runSeurat}
\title{Run Seurat}
\usage{
runSeurat(object, virtualenv = "r-reticulate", regressCellCycle = TRUE)
}
\description{
Run Seurat
}
\note{
Updated 2020-02-20.
}
\section{reticulate}{


Check \code{Sys.getenv("WORKON_HOME")} path, which is the current approach used
in \code{Rprofile.site}.
}

\seealso{
\itemize{
\item https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html
\item https://github.com/satijalab/seurat/wiki
}
}
