% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runSeurat-methods.R
\name{runSeurat}
\alias{runSeurat}
\alias{runSeurat,Seurat-method}
\alias{runSeurat,SingleCellExperiment-method}
\title{Run Seurat}
\usage{
runSeurat(object, ...)

\S4method{runSeurat}{Seurat}(
  object,
  regressCellCycle = c("s-g2m-diff", "yes", "no"),
  varsToRegress = c("nCount_RNA", "mitoRatio"),
  dims = "auto",
  resolution = seq(from = 0.2, to = 1.2, by = 0.2),
  tsneMethod = "Rtsne",
  umapMethod = "umap-learn",
  virtualenv = "r-reticulate",
  workers = "auto"
)

\S4method{runSeurat}{SingleCellExperiment}(object, ...)
}
\description{
Run Seurat
}
\note{
Updated 2020-02-20.
}
\section{reticulate}{


Check \code{Sys.getenv("WORKON_HOME")} path, which is the current approach used
in \code{Rprofile.site}.
}

\seealso{
\itemize{
\item https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html
\item https://github.com/satijalab/seurat/wiki
\item https://satijalab.org/seurat/essential_commands.html
\item https://satijalab.org/seurat/v3.1/cell_cycle_vignette.html
\item Parallelization with future
https://satijalab.org/seurat/v3.0/future_vignette.html
}
}
