% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runZinbwave-methods.R
\docType{methods}
\name{runZinbwave}
\alias{runZinbwave}
\alias{runZinbwave,SingleCellExperiment-method}
\title{Calculate ZINB-WaVE Weights}
\usage{
runZinbwave(Y, ...)

\S4method{runZinbwave}{SingleCellExperiment}(Y, K = 0L,
  epsilon = 1e+12, BPPARAM = BiocParallel::SerialParam(),
  recalculate = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns). Currently implemented
only for \code{SummarizedExperiment}.}

\item{...}{Passthrough arguments to \code{\link[zinbwave:zinbwave]{zinbwave::zinbwave()}}.}

\item{K}{integer. Number of latent factors.}

\item{epsilon}{nonnegative scalar. Regularization parameter.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{\link[BiocParallel]{bpparam}} for details.}

\item{recalculate}{\code{logical}. Force recalculation of weights.}

\item{verbose}{\code{logical}. Run zinbwave in verbose model (for debugging).}
}
\value{
Modified object (S4 class must extend \code{SingleCellExperiment}), with
weights added to \code{\link[=assays]{assays()}}.
}
\description{
zinbwave will calculate \code{normalizedValues} and \code{weights} matrices, which will
be defined in the \code{\link[=assays]{assays()}} slot. If these values have already been
calculated, the object will be returned unmodified, unless
\code{recalculate = TRUE}.
}
\note{
zinbwave defaults to using \code{\link[BiocParallel:bpparam]{BiocParallel::bpparam()}} to register the
number of cores to use (\code{BPPARAM} argument), but we've found that this
works inconsistently across installations. Currently, we recommend using
\code{\link[BiocParallel:SerialParam]{BiocParallel::SerialParam()}} by default, which will run in serial, using a
single core. On Linux or macOS, \code{\link[BiocParallel:MulticoreParam]{BiocParallel::MulticoreParam()}} should
work to run with multiple cores.
}
\examples{
# Example using 200 non-zero genes
Y <- sce_small
Y <- bcbioSingleCell::filterCells(Y)
Y <- Y[seq_len(200L), ]

# Serial
zinb <- runZinbwave(
    Y = Y,
    BPPARAM = BiocParallel::SerialParam()
)
print(zinb)
assayNames(zinb)

# Multicore
zinb <- runZinbwave(
    Y = Y,
    BPPARAM = BiocParallel::MulticoreParam(),
    verbose = TRUE
)
print(zinb)
assayNames(zinb)
}
\seealso{
\code{\link[zinbwave:zinbwave]{zinbwave::zinbwave()}}.

Other Differential Expression Functions: \code{\link{diffExp}},
  \code{\link{findAllMarkers}}
}
\author{
Michael Steinbaugh
}
\concept{Differential Expression Functions}
