% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runZinbwave-methods.R
\name{runZinbwave}
\alias{runZinbwave}
\alias{runZinbwave,SingleCellExperiment-method}
\title{Run ZINB-WaVE}
\usage{
runZinbwave(Y, ...)

\S4method{runZinbwave}{SingleCellExperiment}(
  Y,
  K = 0L,
  epsilon = 1e+12,
  BPPARAM = BiocParallel::bpparam(),
  recalculate = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns).}

\item{...}{Additional arguments.}

\item{K}{integer. Number of latent factors.}

\item{epsilon}{nonnegative scalar. Regularization parameter.}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}

\item{recalculate}{\code{logical}. Force recalculation of weights.}

\item{verbose}{\code{logical}. Run zinbwave in verbose mode (for debugging).}
}
\value{
Modified object.
The dimensionality reduced matrix is stored in the \code{reducedDims} slot,
and optionally normalized values and residuals are added in the list
of assays.
}
\description{
Run ZINB-WaVE
}
\details{
\href{https://bioconductor.org/packages/release/bioc/html/zinbwave.html}{zinbwave} will calculate \code{normalizedValues} and \code{weights} matrices, which
will be defined in the \code{assays} slot. If these values have already been
calculated, the object will be returned unmodified, unless \code{recalculate = TRUE}.
}
\note{
Updated 2021-03-03.
}
\section{Parallelization}{


\code{zinbwave()} defaults to using \code{BiocParallel::bpparam} to register the number
of cores to use (\code{bpparam} argument), but we've found that this works
inconsistently across installations. Currently, we recommend using
\code{BiocParallel::SerialParam} by default, which will run in serial, using a
single core. On Linux or macOS, \code{BiocParallel::MulticoreParam} should work
to run with multiple cores.
}

\examples{
## This example is CPU intensive.
## > if (requireNamespace("zinbwave", quietly = TRUE)) {
## >     data(SingleCellExperiment, package = "AcidTest")
## >     Y <- SingleCellExperiment
## >     Y <- nonzeroRowsAndCols(Y)
## >     Y <- runZinbwave(Y)
## > }
}
\seealso{
\code{zinbwave::zinbwave}.
}
