% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runZinbwave-methods.R
\name{runZinbwave}
\alias{runZinbwave}
\alias{runZinbwave,SingleCellExperiment-method}
\title{Run ZINB-WaVE}
\usage{
runZinbwave(Y, ...)

\S4method{runZinbwave}{SingleCellExperiment}(
  Y,
  K = 0L,
  epsilon = 1e+12,
  BPPARAM = BiocParallel::bpparam(),
  recalculate = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{The data (genes in rows, samples in columns). Currently implemented
only for \code{SummarizedExperiment}.}

\item{...}{Additional parameters to describe the model, see
\code{\link[zinbwave]{zinbModel}}.}

\item{K}{integer. Number of latent factors. Specify \code{K = 0} if only
computing observational weights.}

\item{epsilon}{nonnegative scalar. Regularization parameter.}

\item{BPPARAM}{object of class \code{bpparamClass} that specifies the
back-end to be used for computations. See
\code{\link[BiocParallel]{bpparam}} for details.}

\item{recalculate}{\code{logical}. Force recalculation of weights.}

\item{verbose}{\code{logical}. Run zinbwave in verbose mode (for debugging).}
}
\value{
Modified object.
Weights added to \code{assays}.
}
\description{
Given an object with the data, it performs dimensionality reduction using
a ZINB regression model with gene and cell-level covariates.
}
\details{
\href{https://bioconductor.org/packages/release/bioc/html/zinbwave.html}{zinbwave} will calculate \code{normalizedValues} and \code{weights} matrices, which
will be defined in the \code{assays} slot. If these values have already been
calculated, the object will be returned unmodified, unless \code{recalculate = TRUE}.
}
\note{
Updated 2020-02-21.
}
\section{Parallelization}{


\code{zinbwave()} defaults to using \code{BiocParallel::bpparam} to register the number
of cores to use (\code{bpparam} argument), but we've found that this works
inconsistently across installations. Currently, we recommend using
\code{BiocParallel::SerialParam} by default, which will run in serial, using a
single core. On Linux or macOS, \code{BiocParallel::MulticoreParam} should work
to run with multiple cores.
}

\examples{
data(SingleCellExperiment, package = "acidtest")
Y <- SingleCellExperiment
Y <- nonzeroRowsAndCols(Y)
Y <- runZinbwave(Y)
}
\seealso{
\code{zinbwave::zinbwave}.
}
