% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zzz-seurat-methods.R
\docType{methods}
\name{barcodeRanksPerSample}
\alias{barcodeRanksPerSample}
\alias{cell2sample}
\alias{convertGenesToSymbols}
\alias{gene2symbol}
\alias{interestingGroups}
\alias{interestingGroups<-}
\alias{metrics}
\alias{metricsPerSample}
\alias{plotBarcodeRanks}
\alias{plotCellCounts}
\alias{plotGene}
\alias{plotGenesPerCell}
\alias{plotMitoRatio}
\alias{plotMitoVsCoding}
\alias{plotNovelty}
\alias{plotQC}
\alias{plotUMIsPerCell}
\alias{plotUMIsVsGenes}
\alias{plotViolin}
\alias{plotZerosVsDepth}
\alias{topBarcodes}
\alias{seurat-SingleCellExperiment}
\alias{assay,seurat,ANY-method}
\alias{assayNames,seurat-method}
\alias{assays,seurat-method}
\alias{barcodeRanksPerSample,seurat-method}
\alias{cell2sample,seurat-method}
\alias{cellCountsPerCluster,seurat-method}
\alias{clusterCellCountsPerSample,seurat-method}
\alias{colData,seurat-method}
\alias{colData<-,seurat,DataFrame-method}
\alias{colnames,seurat-method}
\alias{convertGenesToSymbols,seurat-method}
\alias{counts,seurat-method}
\alias{diffExp,seurat-method}
\alias{fetchGeneData,seurat-method}
\alias{fetchReducedDimData,seurat-method}
\alias{fetchReducedDimExpressionData,seurat-method}
\alias{gene2symbol,seurat-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,seurat,character-method}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\alias{metrics,seurat-method}
\alias{metricsPerSample,seurat-method}
\alias{plotBarcodeRanks,seurat-method}
\alias{plotCellCounts,seurat-method}
\alias{plotCellTypesPerCluster,seurat-method}
\alias{plotDot,seurat-method}
\alias{plotFeature,seurat-method}
\alias{plotGene,seurat-method}
\alias{plotGenesPerCell,seurat-method}
\alias{plotKnownMarkersDetected,seurat-method}
\alias{plotMarker,seurat-method}
\alias{plotMitoRatio,seurat-method}
\alias{plotMitoVsCoding,seurat-method}
\alias{plotNovelty,seurat-method}
\alias{plotPCA,seurat-method}
\alias{plotQC,seurat-method}
\alias{plotReducedDim,seurat-method}
\alias{plotTopMarkers,seurat-method}
\alias{plotTSNE,seurat-method}
\alias{plotUMAP,seurat-method}
\alias{plotUMIsPerCell,seurat-method}
\alias{plotUMIsVsGenes,seurat-method}
\alias{plotViolin,seurat-method}
\alias{plotZerosVsDepth,seurat-method}
\alias{reducedDims,seurat-method}
\alias{rowData,seurat-method}
\alias{rownames,seurat-method}
\alias{rowRanges,seurat-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,seurat,DataFrame-method}
\alias{sampleNames,seurat-method}
\alias{topBarcodes,seurat-method}
\title{Extend S4 Methods for \code{seurat} Class}
\usage{
barcodeRanksPerSample(object, ...)

cell2sample(object, ...)

convertGenesToSymbols(object, ...)

gene2symbol(object)

interestingGroups(object, ...)

interestingGroups(object, ...) <- value

metrics(object, ...)

metricsPerSample(object, ...)

plotBarcodeRanks(object, ...)

plotCellCounts(object, ...)

plotGene(object, ...)

plotGenesPerCell(object, ...)

plotMitoRatio(object, ...)

plotMitoVsCoding(object, ...)

plotNovelty(object, ...)

plotQC(object, ...)

plotUMIsPerCell(object, ...)

plotUMIsVsGenes(object, ...)

plotViolin(object, ...)

plotZerosVsDepth(object, ...)

topBarcodes(object, ...)

\S4method{assay}{seurat,ANY}(x, i, ...)

\S4method{assayNames}{seurat}(x, ...)

\S4method{assays}{seurat}(x, ..., withDimnames = TRUE)

\S4method{barcodeRanksPerSample}{seurat}(object, ...)

\S4method{cell2sample}{seurat}(object)

\S4method{cellCountsPerCluster}{seurat}(object, interestingGroups)

\S4method{clusterCellCountsPerSample}{seurat}(object)

\S4method{colData}{seurat}(x, ...)

\S4method{colData}{seurat,DataFrame}(x) <- value

\S4method{colnames}{seurat}(x)

\S4method{convertGenesToSymbols}{seurat}(object)

\S4method{counts}{seurat}(object, ...)

\S4method{diffExp}{seurat}(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)

\S4method{fetchGeneData}{seurat}(object, genes, gene2symbol = FALSE)

\S4method{fetchReducedDimData}{seurat}(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups)

\S4method{fetchReducedDimExpressionData}{seurat}(object, genes, reducedDim)

\S4method{gene2symbol}{seurat}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{seurat,character}(object) <- value

\S4method{metadata}{seurat}(x, ...)

\S4method{metadata}{seurat}(x) <- value

\S4method{metrics}{seurat}(object, ...)

\S4method{metricsPerSample}{seurat}(object, f = c("mean", "median",
  "sum"))

\S4method{plotBarcodeRanks}{seurat}(object, ...)

\S4method{plotCellCounts}{seurat}(object, interestingGroups,
  fill = getOption("bcbio.discrete.fill", NULL), title = "cell counts")

\S4method{plotCellTypesPerCluster}{seurat}(object, markers,
  reducedDim = c("TSNE", "UMAP"), expression = c("mean", "median",
  "sum"), headerLevel = 2L, ...)

\S4method{plotDot}{seurat}(object, genes, colMin = -2.5, colMax = 2.5,
  dotMin = 0L, dotScale = 6L,
  color = getOption("bcbio.discrete.color", NULL),
  legend = getOption("bcbio.legend", TRUE))

\S4method{plotFeature}{seurat}(object, features, reducedDim = c("TSNE",
  "UMAP"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", FALSE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L))

\S4method{plotGene}{seurat}(object, genes, geom = c("dot", "violin"),
  legend = getOption("bcbio.legend", TRUE))

\S4method{plotGenesPerCell}{seurat}(object, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), interestingGroups,
  min = 0L, max = Inf, trans = "log2",
  fill = getOption("bcbio.discrete.fill", NULL),
  title = "genes per cell")

\S4method{plotKnownMarkersDetected}{seurat}(object, markers,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotMarker}{seurat}(object, genes, reducedDim = c("TSNE",
  "UMAP"), expression = c("mean", "median", "sum"),
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

\S4method{plotMitoRatio}{seurat}(object, geom = c("violin", "ridgeline",
  "ecdf", "histogram", "boxplot"), interestingGroups, max = 1L,
  fill = getOption("bcbio.discrete.fill", NULL), trans = "sqrt",
  title = "mito ratio")

\S4method{plotMitoVsCoding}{seurat}(object, interestingGroups,
  trendline = FALSE, color = getOption("bcbio.discrete.color", NULL),
  trans = "log2", title = "mito vs. coding")

\S4method{plotNovelty}{seurat}(object, geom = c("violin", "ridgeline",
  "ecdf", "histogram", "boxplot"), interestingGroups, min = 0L,
  fill = getOption("bcbio.discrete.fill", NULL), trans = "identity",
  title = "genes per UMI (novelty)")

\S4method{plotPCA}{seurat}(object, ...)

\S4method{plotQC}{seurat}(object, interestingGroups, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), headerLevel = 2L,
  legend = getOption("bcbio.legend", FALSE), return = c("grid", "list",
  "markdown"))

\S4method{plotReducedDim}{seurat}(object, reducedDim, dimsUse = c(1L,
  2L), interestingGroups = "ident",
  color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.75),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = NULL)

\S4method{plotTopMarkers}{seurat}(object, markers, n = 10L,
  direction = c("positive", "negative", "both"), coding = FALSE,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

\S4method{plotTSNE}{seurat}(object, ...)

\S4method{plotUMAP}{seurat}(object, ...)

\S4method{plotUMIsPerCell}{seurat}(object, geom = c("violin",
  "ridgeline", "ecdf", "histogram", "boxplot"), interestingGroups,
  min = 0L, max = Inf, point = c("none", "inflection", "knee"),
  trans = "log10", color = getOption("bcbio.discrete.color", NULL),
  fill = getOption("bcbio.discrete.fill", NULL),
  title = "UMIs per cell")

\S4method{plotUMIsVsGenes}{seurat}(object, interestingGroups,
  trendline = FALSE, color = getOption("bcbio.discrete.color", NULL),
  trans = "log2", title = "UMIs vs. genes")

\S4method{plotViolin}{seurat}(object, genes, scale = c("count", "width",
  "area"), fill = getOption("bcbio.discrete.fill", NULL),
  legend = getOption("bcbio.legend", TRUE))

\S4method{plotZerosVsDepth}{seurat}(object, interestingGroups,
  color = getOption("bcbio.discrete.color", NULL),
  title = "zeros vs. depth")

\S4method{reducedDims}{seurat}(x)

\S4method{rowData}{seurat}(x, ...)

\S4method{rownames}{seurat}(x)

\S4method{rowRanges}{seurat}(x)

\S4method{sampleData}{seurat}(object, interestingGroups)

\S4method{sampleData}{seurat,DataFrame}(object) <- value

\S4method{sampleNames}{seurat}(object)

\S4method{topBarcodes}{seurat}(object, n = 1000L,
  return = c("data.frame", "list"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{value}{Object to assign.}

\item{x}{Primary object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}

\item{genes}{\code{character}. Gene identifiers. Must match the rownames of the
object.}

\item{gene2symbol}{\code{data.frame}. Gene-to-symbol mappings. Columns must
contain \code{geneID} and \code{geneName}.}

\item{reducedDim}{\code{string}. Name of reduced dimension matrix slotted in
\code{\link[=reducedDims]{reducedDims()}}. Includes TNSE, UMAP, PCA, for example.}

\item{dimsUse}{\code{integer}. Vector of length 2 that denotes the columns from
the reduced dimension matrix to use for \code{centerX} and \code{centerY} column
calculations. Defaults the first and second dimensions.}

\item{fill}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 fill scale.
Must supply discrete values. When set to \code{NULL}, the default ggplot2 color
palette will be used. If manual color definitions are desired, we recommend
using \code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_fill_manual()}}.
To set the discrete fill palette globally, use
\code{options(bcbio.discrete.fill = scale_fill_viridis_d())}.}

\item{title}{\code{string} or \code{NULL}. Plot title.}

\item{expression}{\code{string}. Calculation to apply. Uses \code{\link[=match.arg]{match.arg()}} and
defaults to the first argument in the \code{character} vector.}

\item{headerLevel}{\code{scalar integer} (\code{1}-\code{7}). R Markdown header level.}

\item{color}{\code{ggproto}/\code{ScaleDiscrete} or \code{NULL}. Desired ggplot2 color
scale. Must supply discrete values. When set to \code{NULL}, the default ggplot2
color palette will be used. If manual color definitions are desired, we
recommend using \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.
To set the discrete color palette globally, use
\code{options(bcbio.discrete.color = scale_color_viridis_d())}.}

\item{legend}{\code{boolean}. Include plot legend.}

\item{pointSize}{\code{scalar numeric}. Cell point size.}

\item{pointAlpha}{\code{scalar numeric} (\code{0}-\code{1}). Alpha transparency level.
Useful when there many cells in the dataset, and some cells can be masked.}

\item{label}{\code{boolean}. Overlay a cluster identitiy label on the plot.}

\item{labelSize}{\code{scalar integer}. Size of the text label.}

\item{dark}{\code{boolean}. Plot against a dark background using
\code{\link[basejump:theme_midnight]{basejump::theme_midnight()}}.}

\item{grid}{\code{boolean}. Show major grid lines but hide axis lines.}

\item{aspectRatio}{\code{scalar integer}. Aspect ratio.}

\item{geom}{\code{string}. Plot type. Uses \code{\link[=match.arg]{match.arg()}} and defaults to the first
argument in the \code{character} vector.}

\item{min}{\code{scalar numeric}. Recommended minimum value cutoff.}

\item{max}{\code{scalar numeric}. Recommended maximum value cutoff.}

\item{trans}{\code{string}. Name of the axis scale transformation to apply. See
\code{help("scale_x_continuous", "ggplot2")} for more information.}

\item{pointsAsNumbers}{\code{boolean}. Plot the points as numbers (\code{TRUE}) or
dots (\code{FALSE}).}

\item{trendline}{\code{boolean}. Include trendline on plot.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
