% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-SingleCellExperiment-methods.R
\docType{methods}
\name{seurat-SingleCellExperiment}
\alias{seurat-SingleCellExperiment}
\alias{assay,seurat,ANY-method}
\alias{assayNames,seurat-method}
\alias{assays,seurat-method}
\alias{colData,seurat-method}
\alias{colData<-,seurat,DataFrame-method}
\alias{colnames,seurat-method}
\alias{counts,seurat-method}
\alias{gene2symbol,seurat-method}
\alias{interestingGroups,seurat-method}
\alias{interestingGroups<-,seurat,character-method}
\alias{metadata,seurat-method}
\alias{metadata<-,seurat-method}
\alias{metrics,seurat-method}
\alias{reducedDims,seurat-method}
\alias{rowData,seurat-method}
\alias{rownames,seurat-method}
\alias{rowRanges,seurat-method}
\alias{sampleData,seurat-method}
\alias{sampleData<-,seurat,DataFrame-method}
\alias{sampleNames,seurat-method}
\title{Extend S4 Methods for \code{seurat} Class}
\usage{
\S4method{assay}{seurat,ANY}(x, i, ...)

\S4method{assayNames}{seurat}(x, ...)

\S4method{assays}{seurat}(x, ..., withDimnames = TRUE)

\S4method{colData}{seurat}(x, ...)

\S4method{colData}{seurat,DataFrame}(x) <- value

\S4method{colnames}{seurat}(x)

\S4method{counts}{seurat}(object, ...)

\S4method{gene2symbol}{seurat}(object)

\S4method{interestingGroups}{seurat}(object)

\S4method{interestingGroups}{seurat,character}(object) <- value

\S4method{metadata}{seurat}(x, ...)

\S4method{metadata}{seurat}(x) <- value

\S4method{metrics}{seurat}(object, ...)

\S4method{reducedDims}{seurat}(x)

\S4method{rowData}{seurat}(x, ...)

\S4method{rownames}{seurat}(x)

\S4method{rowRanges}{seurat}(x)

\S4method{sampleData}{seurat}(object, interestingGroups = NULL)

\S4method{sampleData}{seurat,DataFrame}(object) <- value

\S4method{sampleNames}{seurat}(object)
}
\arguments{
\item{x}{Primary object.}

\item{...}{Additional arguments.}

\item{value}{Object to assign.}

\item{object}{Object.}

\item{interestingGroups}{\code{character} or \code{NULL}. Character vector of
interesting groups. Must be formatted in camel case and intersect with
\code{\link[=sampleData]{sampleData()}} colnames.}
}
\value{
Match \code{SummarizedExperiment} method return.
}
\description{
Provide limited \code{SingleCellExperiment}-like method support.
}
\seealso{
\code{getMethod("metadata<-", "Annotated")}
}
\author{
Michael Steinbaugh
}
\keyword{internal}
