% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seurat}
\alias{Seurat}
\title{Seurat}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
data(Seurat)
}
\value{
\code{Seurat}.
}
\description{
Example \code{Seurat} (v4) object.
}
\note{
Updated 2022-03-10.
}
\examples{
data(Seurat)
print(Seurat)
}
\keyword{datasets}
