% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat}
\alias{seurat}
\title{Seurat}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
data(seurat)
}
\value{
\code{Seurat}.
}
\description{
Example \code{Seurat} (v4) object.
}
\note{
Updated 2022-05-11.
}
\examples{
data(seurat)
class(seurat)
}
\keyword{datasets}
