% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_all_markers}
\alias{seurat_all_markers}
\title{Seurat all markers example}
\format{
An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 76 x 37 x 8 x 8 x 8.
}
\usage{
data(seurat_all_markers)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Seurat all markers example
}
\note{
Updated 2020-01-30.
}
\examples{
data(seurat_all_markers)
seurat_all_markers
}
\keyword{datasets}
