% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_small}
\alias{seurat_small}
\title{Seurat example}
\format{An object of class \code{Seurat} with 230 rows and 80 columns.}
\usage{
seurat_small
}
\description{
Seurat example
}
\examples{
data(seurat_small)
seurat_small
}
\keyword{datasets}
