% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratAllMarkers}
\alias{seuratAllMarkers}
\title{Seurat all markers example}
\format{
An object of class \code{SeuratMarkersPerCluster} of dimension 106 x 79 x 37 x 8 x 8 x 8.
}
\usage{
data(seuratAllMarkers)
}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Seurat all markers example
}
\note{
Updated 2021-03-03.
}
\examples{
data(seuratAllMarkers)
seuratAllMarkers
}
\keyword{datasets}
