% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratKnownMarkers}
\alias{seuratKnownMarkers}
\title{Seurat known markers example}
\format{
An object of class \code{KnownMarkers} with 10 rows and 10 columns.
}
\usage{
data(seuratKnownMarkers)
}
\value{
\code{KnownMarkers}.
}
\description{
Seurat known markers example
}
\note{
Updated 2021-03-03.
}
\examples{
data(seuratKnownMarkers)
seuratKnownMarkers
}
\keyword{datasets}
