% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkers-class}
\alias{SeuratMarkers-class}
\title{\code{SeuratMarkers} Class}
\value{
\code{SeuratMarkers}.
}
\description{
Class containing essential elements for Seurat marker analysis.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{DataFrame}. Sanitized Seurat markers data.}

\item{\code{GRanges}}{\code{GRanges}.}

\item{\code{organism}}{\code{string}. Full Latin organism name.}

\item{\code{ensemblRelease}}{\code{scalar integer}. Ensembl release version.}

\item{\code{version}}{\code{package_version}.}

\item{\code{date}}{\code{Date}. Date the object was saved.}

\item{\code{sessionInfo}}{\code{session_info}. \code{\link[sessioninfo:session_info]{sessioninfo::session_info()}} return.}
}}

\seealso{
Other S4 Classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
