% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkers-class}
\alias{SeuratMarkers-class}
\title{Seurat markers}
\value{
\code{SeuratMarkers}.
}
\description{
Class containing essential elements for marker gene analysis.
}
\details{
Results are arranged by adjusted \emph{P} value (\code{padj}).
}
\note{
Updated 2019-09-03.
}
