% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkers-class}
\alias{SeuratMarkers-class}
\title{\code{SeuratMarkers} Class}
\description{
Class containing essential elements for Seurat marker gene analysis.
}
\seealso{
\code{\link[=SeuratMarkers]{SeuratMarkers()}}.

Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownSeuratMarkers-class}}
}
\concept{S4 classes}
