% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkers-class}
\alias{SeuratMarkers-class}
\title{\code{SeuratMarkers} Class}
\description{
Class containing essential elements for Seurat marker analysis.
}
\seealso{
Other S4 Classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Classes}
