% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{SeuratMarkers-class}
\alias{SeuratMarkers-class}
\alias{SeuratMarkers}
\title{Seurat markers}
\usage{
SeuratMarkers(object, ...)
}
\value{
\code{SeuratMarkers}.
Results are arranged by adjusted \emph{P} value (\code{padj}).
}
\description{
Class containing essential elements for marker gene analysis.
}
\note{
Updated 2019-08-06.
}
