% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkers.R
\name{SeuratMarkers}
\alias{SeuratMarkers}
\title{Sanitize Seurat Markers}
\usage{
SeuratMarkers(markers, ranges, alpha = 0.05)
}
\arguments{
\item{markers}{\code{data.frame}. Unmodified \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} or
\code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} return.}

\item{ranges}{\code{GRanges}. Gene annotations. Names must correspond to the
rownames defined in \code{seurat@data}. The function will automatically subset
the ranges and arrange them alphabetically.}

\item{alpha}{\code{scalar numeric}. Alpha cutoff (adjusted P value; false
discovery rate).}
}
\value{
\code{SeuratMarkers}. Results are arranged by adjusted \emph{P} value, and
grouped per cluster if applicable.
}
\description{
This generator function is designed to take the original return from a Seurat
marker analysis and add corresponding gene annotations.
}
\note{
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} maps the counts matrix rownames correctly in
the \code{gene} column, whereas \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} maps them correctly in
the rownames of the returned marker \code{data.frame}.
}
\examples{
data(seurat_small)
object <- seurat_small

## `FindAllMarkers()` return.
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(object)
))
all_sanitized <- SeuratMarkers(
    markers = all_markers,
    ranges = rowRanges(object)
)
summary(all_sanitized)

## `FindMarkers()` return.
invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "1",
        ident.2 = NULL
    )
))
ident_3_sanitized <- SeuratMarkers(
    markers = ident_3_markers,
    ranges = rowRanges(object)
)
summary(ident_3_sanitized)
}
