% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkers.R, R/SeuratMarkersPerCluster.R
\name{SeuratMarkers}
\alias{SeuratMarkers}
\alias{SeuratMarkersPerCluster}
\title{Seurat Markers}
\usage{
SeuratMarkers(object, ranges, alpha = 0.05)

SeuratMarkersPerCluster(object, ranges, alpha = 0.05)
}
\arguments{
\item{object}{\code{data.frame}. Unmodified \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} or
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{ranges}{\code{GRanges}. Gene annotations. Names must correspond to the
rownames. The function will automatically subset the ranges and arrange
them alphabetically.}

\item{alpha}{\code{scalar numeric}. Alpha cutoff (adjusted P value; false
discovery rate).}
}
\value{
Results are arranged by adjusted \emph{P} value (\code{padj}).
}
\description{
This generator function is designed to take the original return from a Seurat
marker analysis and add corresponding gene annotations.
}
\note{
For \code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return, rownames are correctly returned
in the \code{gene} column.
}
\examples{
data(seurat_small)
object <- seurat_small
ranges <- rowRanges(object)

## `FindMarkers()` return.
invisible(capture.output(
    markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "1",
        ident.2 = "0"
    )
))
x <- MarkersPerCluster(object = markers, ranges = ranges)
summary(ident_3_sanitized)

## `FindAllMarkers()` return.
invisible(capture.output(
    markers <- Seurat::FindAllMarkers(object)
))
x <- MarkersPerCluster(object = markers, ranges = ranges)
summary(x)
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkersPerCluster-class}}
}
