% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster-class}
\title{Seurat markers per cluster}
\value{
Results are split per \code{cluster} and arranged by adjusted \emph{P} value
(\code{padj}).
}
\description{
Class containing essential elements for marker per cluster analysis.
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 classes}
