% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster-class}
\title{Seurat markers per cluster}
\value{
\code{SeuratMarkersPerCluster}.
}
\description{
Class containing essential elements for marker per cluster analysis.
}
\details{
Results are split per \code{cluster} and arranged by adjusted \emph{P} value (\code{padj}).
}
\note{
Updated 2021-03-03.
}
