% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R
\docType{class}
\name{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster}
\title{Seurat markers per cluster}
\usage{
SeuratMarkersPerCluster(object, ...)
}
\value{
\code{SeuratMarkersPerCluster}.
Results are split per \code{cluster} and arranged by adjusted \emph{P} value
(\code{padj}).
}
\description{
Class containing essential elements for marker per cluster analysis.
}
\note{
Updated 2019-08-06.
}
