% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SeuratMarkers.R
\docType{class}
\name{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster-class}
\alias{SeuratMarkersPerCluster}
\title{Seurat markers per cluster}
\usage{
SeuratMarkersPerCluster(object, ranges, alpha = 0.05)
}
\arguments{
\item{object}{\code{data.frame}. Unmodified \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}} or
\code{\link[Seurat:FindAllMarkers]{Seurat::FindAllMarkers()}} return.}

\item{ranges}{\code{GRanges}. Gene annotations. Names must correspond to the
rownames. The function will automatically subset the ranges and arrange
them alphabetically.}

\item{alpha}{\code{numeric(1)}.
Adjusted P value ("alpha") cutoff. If left \code{NULL}, will use the cutoff
defined in the object.}
}
\value{
\code{SeuratMarkersPerCluster}.
Results are split per \code{cluster} and arranged by adjusted \emph{P} value
(\code{padj}).
}
\description{
Class containing essential elements for marker per cluster analysis.
}
\note{
Updated 2019-07-31.
}
\examples{
data(seurat)
object <- seurat
ranges <- rowRanges(object)

## `FindMarkers()` return.
invisible(capture.output(
    markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "1",
        ident.2 = NULL
    )
))
x <- SeuratMarkers(object = markers, ranges = ranges)
summary(x)

## `FindAllMarkers()` return.
invisible(capture.output(suppressWarnings(
    markers <- Seurat::FindAllMarkers(object)
)))
x <- SeuratMarkersPerCluster(object = markers, ranges = ranges)
summary(x)
}
\seealso{
Other S4 classes: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}},
  \code{\link{KnownMarkers-class}},
  \code{\link{SeuratMarkers-class}}
}
\concept{S4 classes}
