% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show}
\alias{show}
\alias{show,CellCycleMarkers-method}
\alias{show,CellTypeMarkers-method}
\title{Show an object}
\usage{
\S4method{show}{CellCycleMarkers}(object)

\S4method{show}{CellTypeMarkers}(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
No return.
}
\description{
Display the object, by printing, plotting or whatever suits its
  class.  This function exists to be specialized by methods.  The
  default method calls \code{\link[methods]{showDefault}}.

  Formal methods for \code{show} will
  usually be invoked for automatic printing (see the details).
}
\note{
Updated 2020-02-20.
}
\examples{
data(cell_cycle_markers_list, cellTypeMarkersList)

## CellCycleMarkers ====
object <- cell_cycle_markers_list[[1L]]
show(object)

## CellTypeMarkers ====
object <- cellTypeMarkersList[[1L]]
show(object)
}
\author{
Michael Steinbuagh
}
