% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SingleCellMarkers-class}
\alias{SingleCellMarkers-class}
\title{\code{SingleCellMarkers} Class}
\description{
Class containing essential elements generated during differential expression
analysis with either Seurat, edgeR, or DESeq2. This class is essentially a
\code{list} with validity checks to ensure that the slotted \code{DataFrame} and
\code{GRanges} correspond.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{grouped_df}.}

\item{\code{rowRanges}}{\code{GRanges}.}
}}

\seealso{
Other S4 Object: \code{\link{CellCycleMarkers-class}},
  \code{\link{CellTypeMarkers-class}}, \code{\link{coerce}}
}
\author{
Michael Steinbaugh
}
\concept{S4 Object}
