% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactors-methods.R
\docType{methods}
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors,SingleCellExperiment-method}
\alias{sizeFactors,cell_data_set-method}
\title{Size factors}
\usage{
\S4method{sizeFactors}{SingleCellExperiment}(object)

\S4method{sizeFactors}{cell_data_set}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{numeric}.
Named numeric vector, corresponding to cells.
}
\description{
Size factors
}
\examples{
data(SingleCellExperiment, package = "acidtest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
}
