% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFactors-methods.R
\docType{methods}
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{sizeFactors,cell_data_set-method}
\alias{sizeFactors<-,cell_data_set,ANY-method}
\title{Size factors}
\usage{
sizeFactors(object, ...)

sizeFactors(object, ...) <- value

\S4method{sizeFactors}{cell_data_set}(object)

\S4method{sizeFactors}{cell_data_set,ANY}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
Names correspond to object column names.
}
\description{
Size factors
}
\note{
For \code{Seurat} objects, use the \code{Seurat::NormalizeData()} function or the
sctransform package to normalize counts, rather than using size factors.

Updated 2019-08-06.
}
\examples{
data(
    SingleCellExperiment,
    cell_data_set,
    package = "acidtest"
)

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
mean(sizeFactors(object))

## cell_data_set ====
object <- cell_data_set
object <- estimateSizeFactors(object)
head(sizeFactors(object))
mean(sizeFactors(object))
}
\seealso{
\itemize{
\item \code{monocle3::size_factors()}.
}
}
