% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,SeuratMarkers-method}
\alias{summary,SeuratMarkersPerCluster-method}
\title{Object summaries}
\usage{
summary(object, ...)

\S4method{summary}{SeuratMarkers}(object)

\S4method{summary}{SeuratMarkersPerCluster}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
\code{summary} is a generic function used to produce result summaries
  of the results of various model fitting functions.  The function
  invokes particular \code{\link{methods}} which depend on the
  \code{\link{class}} of the first argument.
}
\examples{
data(rse, txse, package = "acidtest")

## Gene2Symbol ====
x <- Gene2Symbol(rse)
summary(x)

## Tx2Gene ====
x <- Tx2Gene(txse)
summary(x)
}
\references{
Chambers, J. M. and Hastie, T. J. (1992)
  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
