% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary-methods.R
\name{summary}
\alias{summary}
\alias{summary,SeuratMarkers-method}
\alias{summary,SeuratMarkersPerCluster-method}
\title{Object Summaries}
\usage{
summary(object, ...)

\S4method{summary}{SeuratMarkers}(object)

\S4method{summary}{SeuratMarkersPerCluster}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
The form of the value returned by \code{summary} depends on the
    class of its argument.  See the documentation of the particular
    methods for details of what is produced by that method.

    The default method returns an object of class
    \code{c("summaryDefault", "\link[base]{table}")} which has specialized
    \code{\link[base]{format}} and \code{\link[base]{print}} methods.  The
    \code{\link[base]{factor}} method returns an integer vector.

    The matrix and data frame methods return a matrix of class
    \code{"\link[base]{table}"}, obtained by applying \code{summary} to each
    column and collating the results.
}
\description{
\code{summary} is a generic function used to produce result summaries
  of the results of various model fitting functions.  The function
  invokes particular \code{\link{methods}} which depend on the
  \code{\link[base]{class}} of the first argument.
}
\note{
Updated 2022-05-11.
}
\examples{
data(SeuratMarkersPerCluster)

## SeuratMarkersPerCluster ====
object <- SeuratMarkersPerCluster
class(object)
summary(object)
}
