% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topMarkers-methods.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,SeuratMarkers-method}
\title{Top Markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{SeuratMarkers}(object, n = 10L,
  direction = c("up", "down", "both"), return = c("tbl_df",
  "DataFrame", "DataFrameList"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{n}{\code{scalar integer}. Number of genes per cluster.}

\item{direction}{\code{string}. Whether to include upregulated (\code{"up"}; positive
LFC), downregulated (\code{"down"}; negative LFC) or \code{"both"} directions of
association per cluster.}

\item{return}{\code{string}. Return type. Uses \code{\link[base:match.arg]{base::match.arg()}} internally and
defaults to the first argument in the \code{character} vector.}
}
\value{
\code{grouped_df}.
}
\description{
Top Markers
}
\examples{
data(all_markers_small)
x <- topMarkers(all_markers_small, n = 2L)
print(x)
}
\seealso{
\itemize{
\item \code{\link[dplyr:slice]{dplyr::slice()}}.
\item \code{\link[dplyr:top_n]{dplyr::top_n()}}.
}
}
