% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topMarkers-methods.R
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,SeuratMarkersPerCluster-method}
\title{Top markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{SeuratMarkersPerCluster}(object, n = 10L, direction = c("up", "down", "both"))
}
\arguments{
\item{object}{Object.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Top markers
}
\note{
Updated 2020-01-30.
}
\examples{
data(seurat_all_markers)

## SeuratMarkersPerCluster ====
object <- seurat_all_markers
x <- topMarkers(object, n = 2L)
print(x)
}
