% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topMarkers-methods.R
\docType{methods}
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,SeuratMarkersPerCluster-method}
\title{Top markers}
\usage{
\S4method{topMarkers}{SeuratMarkersPerCluster}(object, n = 10L,
  direction = c("up", "down", "both"), return = c("tbl_df",
  "DataFrame", "SplitDataFrameList"))
}
\arguments{
\item{object}{Object.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{return}{\code{character(1)}.
Return type.
Uses \link[base:match.arg]{match.arg()} internally and defaults to the first
argument in the \code{character} vector.}
}
\value{
\code{grouped_df}.
}
\description{
Top markers
}
\examples{
data(all_markers_small)
x <- topMarkers(all_markers_small, n = 2L)
print(x)
}
\seealso{
\itemize{
\item \code{dplyr::slice}.
\item \code{dplyr::top_n}.
}
}
