% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topMarkers-methods.R
\name{topMarkers}
\alias{topMarkers}
\alias{topMarkers,SeuratMarkersPerCluster-method}
\title{Top markers}
\usage{
topMarkers(object, ...)

\S4method{topMarkers}{SeuratMarkersPerCluster}(object, direction = c("both", "up", "down"), n = 10L)
}
\arguments{
\item{object}{Object.}

\item{direction}{\code{character(1)}.
Whether to include upregulated (\code{"up"}; positive LFC), downregulated
(\code{"down"}; negative LFC) or \code{"both"} directions of association per cluster.}

\item{n}{\code{integer(1)}.
Number of genes per cluster.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Top markers
}
\note{
Updated 2022-05-11.
}
\examples{
data(SeuratMarkersPerCluster)

## SeuratMarkersPerCluster ====
object <- SeuratMarkersPerCluster
x <- topMarkers(
    object = object,
    direction = "up",
    n = 2L
)
print(x)
}
